separate runtime indexing owners
