+++ crates/codestory-runtime/src/lib.rs
@@ -88,7 +88,11 @@ use uuid::Uuid;
 
 mod affected;
 mod agent;
+mod index_commit;
 mod index_freshness;
+mod index_full;
+mod index_incremental;
+mod index_timings;
 mod publication;
 mod repo_text;
 mod route_coverage;
@@ -97,12 +101,15 @@ mod search_plan;
 mod search_publication;
 mod search_scoring;
 mod search_state;
+mod search_state_cache;
 mod search_terms;
 mod semantic_projection;
+mod semantic_republish;
 mod snippets;
 mod workspace_state;
 use affected::{AffectedOperationIdentityIndex, IndexFreshnessObservation};
 pub use agent::{packet_step_trace_json, plan_packet};
+use index_commit::*;
 use index_freshness::{
     CachedIndexFreshness, index_freshness_cache_ttl_secs, index_freshness_from_storage_with_policy,
     index_freshness_observation_from_storage_with_identities, indexable_source_path_in_workspace,
@@ -114,6 +121,9 @@ use index_freshness::{
     EXACT_SYMBOL_HYBRID_MAX_RESULTS_CAP, arm_after_index_freshness_fence_test_hook,
     index_freshness_from_storage, indexable_source_path, not_checked_index_freshness,
 };
+use index_full::*;
+use index_incremental::*;
+use index_timings::*;
 #[cfg(test)]
 use publication::{
     PUBLICATION_TEST_FAULT, PublicationTestAction, PublicationTestBoundary,
@@ -148,6 +158,7 @@ use search_publication::{
 #[cfg(test)]
 use search_scoring::HybridSearchInstrumentation;
 pub(crate) use search_scoring::HybridSearchScoredHit;
+use search_state_cache::*;
 pub use semantic_projection::SemanticProjectionRepublishOutcome;
 #[cfg(feature = "test-support")]
 #[doc(hidden)]
@@ -182,6 +193,7 @@ use semantic_projection::{
     sort_pending_dense_anchor_inputs, stream_pending_llm_symbol_docs_from_env,
     truncate_semantic_doc_text_to_token_budget,
 };
+use semantic_republish::*;
 #[cfg(test)]
 pub(crate) use snippets::markdown_snippet;
 pub(crate) use snippets::{
@@ -212,6 +224,7 @@ mod search_runtime;
 pub mod benchmark_support {
     pub use crate::search::engine::{SearchEngine, SymbolIndexSession, SymbolIndexWriteStats};
 }
+
 mod semantic_doc_text;
 mod services;
 mod support;
@@ -219,6 +232,8 @@ mod symbol_query;
 mod symbol_workflow;
 mod system_actions;
 mod target_resolution;
+#[cfg(test)]
+mod tests;
 mod trail_story;
 
 pub use browser::{BrowserQueryItem, ReadOnlyBrowserService};
@@ -439,247 +454,6 @@ impl Default for Runtime {
 
 fn no_project_error() -> ApiError {
     ApiError::invalid_argument("No project open. Call open_project first.")
-}
-
-#[derive(Debug, Clone, Default)]
-struct OptionalResolutionTelemetry {
-    setup_existing_projection_ids_ms: Option<u32>,
-    setup_seed_symbol_table_ms: Option<u32>,
-    flush_files_ms: Option<u32>,
-    flush_nodes_ms: Option<u32>,
-    flush_edges_ms: Option<u32>,
-    flush_occurrences_ms: Option<u32>,
-    flush_component_access_ms: Option<u32>,
-    flush_callable_projection_ms: Option<u32>,
-    resolution_override_count_ms: Option<u32>,
-    resolution_unresolved_counts_ms: Option<u32>,
-    resolution_calls_ms: Option<u32>,
-    resolution_imports_ms: Option<u32>,
-    resolution_cleanup_ms: Option<u32>,
-    resolution_call_candidate_index_ms: Option<u32>,
-    resolution_import_candidate_index_ms: Option<u32>,
-    resolution_call_semantic_index_ms: Option<u32>,
-    resolution_import_semantic_index_ms: Option<u32>,
-    resolution_support_snapshot_limit_bytes: Option<u64>,
-    resolution_support_snapshot_stored: Option<bool>,
-    resolution_support_snapshot_skipped_oversize: Option<bool>,
-    resolution_call_semantic_candidates_ms: Option<u32>,
-    resolution_import_semantic_candidates_ms: Option<u32>,
-    resolution_call_semantic_requests: Option<u32>,
-    resolution_call_semantic_unique_requests: Option<u32>,
-    resolution_call_semantic_skipped_requests: Option<u32>,
-    resolution_import_semantic_requests: Option<u32>,
-    resolution_import_semantic_unique_requests: Option<u32>,
-    resolution_import_semantic_skipped_requests: Option<u32>,
-    resolution_call_compute_ms: Option<u32>,
-    resolution_import_compute_ms: Option<u32>,
-    resolution_call_apply_ms: Option<u32>,
-    resolution_import_apply_ms: Option<u32>,
-    resolution_override_resolution_ms: Option<u32>,
-    resolved_calls_same_file: Option<u32>,
-    resolved_calls_same_module: Option<u32>,
-    resolved_calls_global_unique: Option<u32>,
-    resolved_calls_semantic: Option<u32>,
-    resolved_imports_same_file: Option<u32>,
-    resolved_imports_same_module: Option<u32>,
-    resolved_imports_global_unique: Option<u32>,
-    resolved_imports_fuzzy: Option<u32>,
-    resolved_imports_semantic: Option<u32>,
-}
-
-fn artifact_cache_access_timings(stats: &ArtifactCacheFamilyStats) -> ArtifactCacheAccessTimings {
-    ArtifactCacheAccessTimings {
-        policy: match stats.policy {
-            ArtifactCachePolicy::KnownEmpty => ArtifactCachePolicyDto::KnownEmpty,
-            ArtifactCachePolicy::ReadThrough => ArtifactCachePolicyDto::ReadThrough,
-        },
-        logical_lookups: clamp_usize_to_u32(stats.logical_lookups),
-        physical_queries: clamp_usize_to_u32(stats.physical_queries),
-        hits: clamp_usize_to_u32(stats.hits),
-        misses: clamp_usize_to_u32(stats.misses),
-        reader_opens: clamp_usize_to_u32(stats.reader_opens),
-        lookup_wall_ms: clamp_u64_to_u32(stats.lookup_wall_ns / 1_000_000),
-    }
-}
-
-fn projection_persistence_family_timings(
-    stats: codestory_store::ProjectionPersistenceFamilyStats,
-) -> ProjectionPersistenceFamilyTimings {
-    ProjectionPersistenceFamilyTimings {
-        row_attempts: stats.row_attempts,
-        bound_bytes: stats.bound_bytes,
-        statement_executions: stats.statement_executions,
-        wall_ms: stats.wall_ms,
-    }
-}
-
-fn projection_persistence_timings(
-    stats: &codestory_store::ProjectionPersistenceStats,
-) -> ProjectionPersistenceTimings {
-    ProjectionPersistenceTimings {
-        transactions: stats.transactions.min(u32::MAX as u64) as u32,
-        row_attempts: stats.row_attempts(),
-        bound_bytes: stats.bound_bytes(),
-        statement_executions: stats.statement_executions(),
-        transaction_wall_ms: stats.transaction_wall_ms,
-        transaction_setup_ms: stats.transaction_setup_ms,
-        commit_ms: stats.commit_ms,
-        files: projection_persistence_family_timings(stats.files),
-        nodes: projection_persistence_family_timings(stats.nodes),
-        structural_text: projection_persistence_family_timings(stats.structural_text),
-        edges: projection_persistence_family_timings(stats.edges),
-        occurrences: projection_persistence_family_timings(stats.occurrences),
-        component_access: projection_persistence_family_timings(stats.component_access),
-        callable_projection: projection_persistence_family_timings(stats.callable_projection),
-        file_errors: projection_persistence_family_timings(stats.file_errors),
-        dirty_state: projection_persistence_family_timings(stats.dirty_state),
-    }
-}
-
-fn database_snapshot_copy_timings(
-    stats: codestory_store::DatabaseSnapshotCopyStats,
-) -> DatabaseSnapshotCopyTimings {
-    DatabaseSnapshotCopyTimings {
-        copy_ms: stats.copy_ms,
-        source_bytes: stats.source_bytes,
-        target_bytes: stats.target_bytes,
-    }
-}
-
-fn core_promotion_timings(stats: codestory_store::CorePromotionStats) -> CorePromotionTimings {
-    CorePromotionTimings {
-        total_ms: stats.total_ms,
-        lock_recovery_ms: stats.lock_recovery_ms,
-        candidate_validation_ms: stats.candidate_validation_ms,
-        previous_validation_ms: stats.previous_validation_ms,
-        rollback_backup_copy_ms: stats.rollback_backup_copy_ms,
-        backup_validation_ms: stats.backup_validation_ms,
-        prepared_journal_write_ms: stats.prepared_journal_write_ms,
-        prepared_journal_file_sync_ms: stats.prepared_journal_file_sync_ms,
-        prepared_journal_directory_sync_ms: stats.prepared_journal_directory_sync_ms,
-        staged_to_live_restore_ms: stats.staged_to_live_restore_ms,
-        promoted_validation_ms: stats.promoted_validation_ms,
-        committed_journal_ms: stats.committed_journal_ms,
-        cleanup_ms: stats.cleanup_ms,
-        unattributed_ms: stats.unattributed_ms,
-        candidate_bytes: stats.candidate_bytes,
-        previous_live_bytes: stats.previous_live_bytes,
-        rollback_backup_bytes: stats.rollback_backup_bytes,
-    }
-}
-
-impl OptionalResolutionTelemetry {
-    fn from_incremental_stats(index_stats: &IncrementalIndexingStats) -> Self {
-        let mut telemetry = Self::from_flush_stats(index_stats);
-        if index_stats.resolution_ran {
-            telemetry.apply_resolution_stats(index_stats);
-        }
-        telemetry
-    }
-
-    fn from_flush_stats(index_stats: &IncrementalIndexingStats) -> Self {
-        Self {
-            setup_existing_projection_ids_ms: Some(clamp_u64_to_u32(
-                index_stats.setup_existing_projection_ids_ms,
-            )),
-            setup_seed_symbol_table_ms: Some(clamp_u64_to_u32(
-                index_stats.setup_seed_symbol_table_ms,
-            )),
-            flush_files_ms: Some(clamp_u64_to_u32(index_stats.flush_files_ms)),
-            flush_nodes_ms: Some(clamp_u64_to_u32(index_stats.flush_nodes_ms)),
-            flush_edges_ms: Some(clamp_u64_to_u32(index_stats.flush_edges_ms)),
-            flush_occurrences_ms: Some(clamp_u64_to_u32(index_stats.flush_occurrences_ms)),
-            flush_component_access_ms: Some(clamp_u64_to_u32(
-                index_stats.flush_component_access_ms,
-            )),
-            flush_callable_projection_ms: Some(clamp_u64_to_u32(
-                index_stats.flush_callable_projection_ms,
-            )),
-            ..Self::default()
-        }
-    }
-
-    fn apply_resolution_stats(&mut self, index_stats: &IncrementalIndexingStats) {
-        self.resolution_override_count_ms =
-            Some(clamp_u64_to_u32(index_stats.resolution_override_count_ms));
-        self.resolution_unresolved_counts_ms = Some(clamp_u64_to_u32(
-            index_stats.resolution_unresolved_counts_ms,
-        ));
-        self.resolution_calls_ms = Some(clamp_u64_to_u32(index_stats.resolution_calls_ms));
-        self.resolution_imports_ms = Some(clamp_u64_to_u32(index_stats.resolution_imports_ms));
-        self.resolution_cleanup_ms = Some(clamp_u64_to_u32(index_stats.resolution_cleanup_ms));
-        self.resolution_call_candidate_index_ms = Some(clamp_u64_to_u32(
-            index_stats.resolution_call_candidate_index_ms,
-        ));
-        self.resolution_import_candidate_index_ms = Some(clamp_u64_to_u32(
-            index_stats.resolution_import_candidate_index_ms,
-        ));
-        self.resolution_call_semantic_index_ms = Some(clamp_u64_to_u32(
-            index_stats.resolution_call_semantic_index_ms,
-        ));
-        self.resolution_import_semantic_index_ms = Some(clamp_u64_to_u32(
-            index_stats.resolution_import_semantic_index_ms,
-        ));
-        self.resolution_support_snapshot_limit_bytes =
-            Some(index_stats.resolution_support_snapshot_limit_bytes);
-        self.resolution_support_snapshot_stored =
-            Some(index_stats.resolution_support_snapshot_stored);
-        self.resolution_support_snapshot_skipped_oversize =
-            Some(index_stats.resolution_support_snapshot_skipped_oversize);
-        self.resolution_call_semantic_candidates_ms = Some(clamp_u64_to_u32(
-            index_stats.resolution_call_semantic_candidates_ms,
-        ));
-        self.resolution_import_semantic_candidates_ms = Some(clamp_u64_to_u32(
-            index_stats.resolution_import_semantic_candidates_ms,
-        ));
-        self.resolution_call_semantic_requests = Some(clamp_usize_to_u32(
-            index_stats.resolution_call_semantic_requests,
-        ));
-        self.resolution_call_semantic_unique_requests = Some(clamp_usize_to_u32(
-            index_stats.resolution_call_semantic_unique_requests,
-        ));
-        self.resolution_call_semantic_skipped_requests = Some(clamp_usize_to_u32(
-            index_stats.resolution_call_semantic_skipped_requests,
-        ));
-        self.resolution_import_semantic_requests = Some(clamp_usize_to_u32(
-            index_stats.resolution_import_semantic_requests,
-        ));
-        self.resolution_import_semantic_unique_requests = Some(clamp_usize_to_u32(
-            index_stats.resolution_import_semantic_unique_requests,
-        ));
-        self.resolution_import_semantic_skipped_requests = Some(clamp_usize_to_u32(
-            index_stats.resolution_import_semantic_skipped_requests,
-        ));
-        self.resolution_call_compute_ms =
-            Some(clamp_u64_to_u32(index_stats.resolution_call_compute_ms));
-        self.resolution_import_compute_ms =
-            Some(clamp_u64_to_u32(index_stats.resolution_import_compute_ms));
-        self.resolution_call_apply_ms =
-            Some(clamp_u64_to_u32(index_stats.resolution_call_apply_ms));
-        self.resolution_import_apply_ms =
-            Some(clamp_u64_to_u32(index_stats.resolution_import_apply_ms));
-        self.resolution_override_resolution_ms = Some(clamp_u64_to_u32(
-            index_stats.resolution_override_resolution_ms,
-        ));
-        self.resolved_calls_same_file =
-            Some(clamp_usize_to_u32(index_stats.resolved_calls_same_file));
-        self.resolved_calls_same_module =
-            Some(clamp_usize_to_u32(index_stats.resolved_calls_same_module));
-        self.resolved_calls_global_unique =
-            Some(clamp_usize_to_u32(index_stats.resolved_calls_global_unique));
-        self.resolved_calls_semantic =
-            Some(clamp_usize_to_u32(index_stats.resolved_calls_semantic));
-        self.resolved_imports_same_file =
-            Some(clamp_usize_to_u32(index_stats.resolved_imports_same_file));
-        self.resolved_imports_same_module =
-            Some(clamp_usize_to_u32(index_stats.resolved_imports_same_module));
-        self.resolved_imports_global_unique = Some(clamp_usize_to_u32(
-            index_stats.resolved_imports_global_unique,
-        ));
-        self.resolved_imports_fuzzy = Some(clamp_usize_to_u32(index_stats.resolved_imports_fuzzy));
-        self.resolved_imports_semantic =
-            Some(clamp_usize_to_u32(index_stats.resolved_imports_semantic));
-    }
 }
 
 fn parse_db_id(raw: &str, field_name: &str) -> Result<i64, ApiError> {
@@ -3651,2596 +3425,3 @@ impl AppController {
         })
     }
 }
-
-fn is_low_confidence_search_plan_bridge(bridge: &SearchPlanBridgeDto) -> bool {
-    bridge.confidence == SearchPlanBridgeConfidenceDto::Low
-}
-
-struct IndexingRunSummary {
-    phase_timings: IndexingPhaseTimings,
-    staged_semantic_stats: SemanticProjectionStats,
-    llm_refresh_scope: Option<HashSet<codestory_contracts::graph::NodeId>>,
-    #[cfg(test)]
-    publication: IndexPublicationRecord,
-    prepared_search_state: Option<SearchStateBuildResult>,
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-struct FullRefreshWallDurations {
-    live_inspection: Duration,
-    source_discovery: Duration,
-    stage_open: Duration,
-    indexer_execution: Duration,
-    coverage_validation: Duration,
-    copy_forward: Duration,
-    semantic_stage: Duration,
-    snapshot_stage: Duration,
-    publication_prepare: Duration,
-    search_generation: Duration,
-    catalog_publication: Duration,
-}
-
-impl FullRefreshWallDurations {
-    fn finish(self, core_refresh: Duration) -> FullRefreshWallTimings {
-        let accounted = [
-            self.live_inspection,
-            self.source_discovery,
-            self.stage_open,
-            self.indexer_execution,
-            self.coverage_validation,
-            self.copy_forward,
-            self.semantic_stage,
-            self.snapshot_stage,
-            self.publication_prepare,
-            self.search_generation,
-            self.catalog_publication,
-        ]
-        .into_iter()
-        .fold(Duration::ZERO, Duration::saturating_add);
-        let unattributed = core_refresh.saturating_sub(accounted);
-
-        FullRefreshWallTimings {
-            core_refresh_ms: clamp_u128_to_u32(core_refresh.as_millis()),
-            live_inspection_ms: clamp_u128_to_u32(self.live_inspection.as_millis()),
-            source_discovery_ms: clamp_u128_to_u32(self.source_discovery.as_millis()),
-            stage_open_ms: clamp_u128_to_u32(self.stage_open.as_millis()),
-            indexer_execution_ms: clamp_u128_to_u32(self.indexer_execution.as_millis()),
-            coverage_validation_ms: clamp_u128_to_u32(self.coverage_validation.as_millis()),
-            copy_forward_ms: clamp_u128_to_u32(self.copy_forward.as_millis()),
-            semantic_stage_ms: clamp_u128_to_u32(self.semantic_stage.as_millis()),
-            snapshot_stage_ms: clamp_u128_to_u32(self.snapshot_stage.as_millis()),
-            publication_prepare_ms: clamp_u128_to_u32(self.publication_prepare.as_millis()),
-            search_generation_ms: clamp_u128_to_u32(self.search_generation.as_millis()),
-            catalog_publication_ms: clamp_u128_to_u32(self.catalog_publication.as_millis()),
-            unattributed_ms: clamp_u128_to_u32(unattributed.as_millis()),
-        }
-    }
-}
-
-fn next_index_publication(
-    previous: Option<&IndexPublicationRecord>,
-    mode: IndexPublicationMode,
-    run_id: &str,
-) -> Result<IndexPublicationRecord, ApiError> {
-    let generation = previous
-        .map(|publication| publication.generation)
-        .unwrap_or_default()
-        .checked_add(1)
-        .ok_or_else(|| ApiError::internal("Index publication generation overflow"))?;
-    Ok(IndexPublicationRecord {
-        generation,
-        generation_id: Uuid::new_v4().to_string(),
-        run_id: run_id.to_string(),
-        mode,
-        published_at_epoch_ms: current_epoch_ms(),
-    })
-}
-
-fn index_publication_dto(publication: IndexPublicationRecord) -> IndexPublicationDto {
-    IndexPublicationDto {
-        generation: publication.generation,
-        generation_id: publication.generation_id,
-        run_id: publication.run_id,
-        mode: match publication.mode {
-            IndexPublicationMode::Full => IndexPublicationModeDto::Full,
-            IndexPublicationMode::Incremental => IndexPublicationModeDto::Incremental,
-            IndexPublicationMode::SemanticProjection => IndexPublicationModeDto::SemanticProjection,
-        },
-        published_at_epoch_ms: publication.published_at_epoch_ms,
-    }
-}
-
-struct IndexWriterGuard {
-    file: std::fs::File,
-    path: PathBuf,
-}
-
-impl IndexWriterGuard {
-    fn try_acquire(storage_path: &Path) -> Result<Self, ApiError> {
-        let path = storage_path.with_extension("index-writer.lock");
-        if let Some(parent) = path
-            .parent()
-            .filter(|parent| !parent.as_os_str().is_empty())
-        {
-            std::fs::create_dir_all(parent).map_err(|error| {
-                ApiError::internal(format!(
-                    "Failed to create index writer lock directory {}: {error}",
-                    parent.display()
-                ))
-            })?;
-        }
-        let file = std::fs::OpenOptions::new()
-            .read(true)
-            .write(true)
-            .create(true)
-            .truncate(false)
-            .open(&path)
-            .map_err(|error| {
-                ApiError::internal(format!(
-                    "Failed to open index writer lock {}: {error}",
-                    path.display()
-                ))
-            })?;
-        if !FileExt::try_lock_exclusive(&file).map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to acquire index writer lock {}: {error}",
-                path.display()
-            ))
-        })? {
-            return Err(ApiError::new(
-                "cache_busy",
-                format!(
-                    "Another indexing run owns the writer lock at {}. Wait for it to finish and retry.",
-                    path.display()
-                ),
-            ));
-        }
-        Ok(Self { file, path })
-    }
-}
-
-impl Drop for IndexWriterGuard {
-    fn drop(&mut self) {
-        if let Err(error) = FileExt::unlock(&self.file) {
-            tracing::warn!(
-                path = %self.path.display(),
-                "Failed to unlock index writer lock: {error}"
-            );
-        }
-    }
-}
-
-fn validate_semantic_projection_core(
-    staged: &mut StagedSnapshot,
-    root: &Path,
-    expected_schema_version: u32,
-    expected_publication: &IndexPublicationRecord,
-    source_index_policy: &SourceIndexPolicy,
-) -> Result<Vec<SourcePolicyExclusionRecord>, ApiError> {
-    let staged_publication = staged
-        .store_mut()
-        .get_complete_index_publication()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to validate the staged core publication: {error}"
-            ))
-        })?
-        .ok_or_else(|| {
-            ApiError::new(
-                "semantic_projection_core_incomplete",
-                "The staged core publication is incomplete.",
-            )
-        })?;
-    if staged_publication != *expected_publication {
-        return Err(ApiError::new(
-            "publication_changed",
-            "The cloned core does not match the pinned live publication.",
-        ));
-    }
-    staged
-        .store_mut()
-        .validate_dense_anchor_publication(expected_publication)
-        .map_err(|error| {
-            ApiError::new(
-                "semantic_projection_migration_required",
-                format!("Pinned dense-anchor publication is not complete: {error}"),
-            )
-        })?;
-    let structural_compatibility = staged
-        .store_mut()
-        .validate_structural_text_unit_publication_or_legacy_empty(expected_publication)
-        .map_err(|error| {
-            ApiError::new(
-                "semantic_projection_migration_required",
-                format!("Pinned structural state is not compatible: {error}"),
-            )
-        })?;
-    if structural_compatibility == StructuralTextPublicationCompatibility::LegacyEmpty
-        && expected_schema_version != LEGACY_SEMANTIC_PROJECTION_SCHEMA_VERSION
-    {
-        return Err(ApiError::new(
-            "semantic_projection_migration_required",
-            "A missing structural publication is compatible only with a schema-29 retained core whose structural stores are empty.",
-        ));
-    }
-    let source_manifest = staged
-        .store_mut()
-        .get_source_policy_exclusion_manifest()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to load the pinned source-policy manifest: {error}"
-            ))
-        })?
-        .ok_or_else(|| {
-            ApiError::new(
-                "semantic_projection_migration_required",
-                "Pinned source-policy publication is missing.",
-            )
-        })?;
-    let recorded_source_policy = SourcePolicyExclusionPolicyIdentity::new(
-        &source_manifest.policy_version,
-        source_manifest.byte_cap,
-        source_manifest.structural_unit_cap,
-    );
-    let selected_identity = project_identity_v3(root);
-    if source_manifest.project_id != selected_identity.project_id
-        || source_manifest.workspace_id != selected_identity.workspace_id
-    {
-        return Err(ApiError::new(
-            "semantic_projection_project_mismatch",
-            "The selected project root does not own the cached core publication.",
-        ));
-    }
-    let compatibility = semantic_projection_source_policy_compatibility(
-        recorded_source_policy,
-        source_index_policy,
-        expected_schema_version,
-        structural_compatibility == StructuralTextPublicationCompatibility::LegacyEmpty,
-    )
-    .ok_or_else(|| {
-        ApiError::new(
-            "semantic_projection_migration_required",
-            "Pinned source-policy identity differs from the current runtime policy; run a source refresh before republishing semantic projections.",
-        )
-    })?;
-    let validation = match compatibility {
-        SemanticProjectionSourcePolicyCompatibility::Exact => staged
-            .store_mut()
-            .validate_source_policy_exclusion_publication(
-                expected_publication,
-                &source_manifest.project_id,
-                &source_manifest.workspace_id,
-                recorded_source_policy,
-            ),
-        SemanticProjectionSourcePolicyCompatibility::LegacyPredecessor => staged
-            .store_mut()
-            .validate_legacy_v1_source_policy_exclusion_publication(
-                expected_publication,
-                &source_manifest.project_id,
-                &source_manifest.workspace_id,
-                recorded_source_policy,
-            ),
-    };
-    validation.map_err(|error| {
-        ApiError::new(
-            "semantic_projection_migration_required",
-            format!("Pinned source-policy publication is not complete: {error}"),
-        )
-    })?;
-    staged
-        .store_mut()
-        .get_source_policy_exclusions()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to load pinned source-policy exclusions: {error}"
-            ))
-        })
-}
-
-struct PreparedSemanticProjection {
-    stats: SemanticProjectionStats,
-    finalize_stats: StagedSnapshotFinalizeStats,
-    detail_snapshot_ms: u32,
-}
-
-fn prepare_semantic_projection(
-    staged: &mut StagedSnapshot,
-    publication: &IndexPublicationRecord,
-    cancel_token: Option<&CancellationToken>,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-) -> Result<PreparedSemanticProjection, ApiError> {
-    let source_identity = format!("core:{}:{}", publication.generation_id, publication.run_id);
-    staged
-        .store_mut()
-        .begin_incremental_run()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to fence the staged semantic projection writer: {error}"
-            ))
-        })?;
-    staged
-        .store_mut()
-        .invalidate_grounding_snapshots()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to invalidate staged derived snapshots: {error}"
-            ))
-        })?;
-
-    let stats = finalize_staged_semantic_docs_for_runtime(
-        staged.store_mut(),
-        None,
-        None,
-        &source_identity,
-        cancel_token,
-        runtime,
-        SemanticProjectionDocumentSource::StoredCore,
-    )?;
-    ensure_indexing_active(cancel_token)?;
-    let finalize_stats = staged.snapshots().finalize_staged().map_err(|error| {
-        ApiError::internal(format!(
-            "Failed to finalize staged semantic projection snapshots: {error}"
-        ))
-    })?;
-    #[cfg(test)]
-    publication_test_checkpoint(
-        PublicationTestBoundary::ProjectionSnapshotFinalize,
-        cancel_token,
-    )?;
-    ensure_indexing_active(cancel_token)?;
-    let detail_started = Instant::now();
-    staged.snapshots().refresh_detail().map_err(|error| {
-        ApiError::internal(format!(
-            "Failed to refresh staged grounding detail snapshot: {error}"
-        ))
-    })?;
-    #[cfg(test)]
-    publication_test_checkpoint(
-        PublicationTestBoundary::ProjectionSnapshotDetail,
-        cancel_token,
-    )?;
-    ensure_indexing_active(cancel_token)?;
-    Ok(PreparedSemanticProjection {
-        stats,
-        finalize_stats,
-        detail_snapshot_ms: clamp_u128_to_u32(detail_started.elapsed().as_millis()),
-    })
-}
-
-fn stage_semantic_projection_publication(
-    staged: &mut StagedSnapshot,
-    root: &Path,
-    publication: &IndexPublicationRecord,
-    source_exclusions: &[SourcePolicyExclusionRecord],
-    source_index_policy: &SourceIndexPolicy,
-    cancel_token: Option<&CancellationToken>,
-) -> Result<u64, ApiError> {
-    #[cfg(test)]
-    publication_test_checkpoint(
-        PublicationTestBoundary::ProjectionManifestIdentity,
-        cancel_token,
-    )?;
-    ensure_indexing_active(cancel_token)?;
-    let dense_manifest = staged
-        .store_mut()
-        .publish_dense_anchor_generation(publication, SEMANTIC_POLICY_VERSION)
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to publish semantic dense-anchor inputs: {error}"
-            ))
-        })?;
-    let selected_identity = project_identity_v3(root);
-    let current_source_policy = SourcePolicyExclusionPolicyIdentity::new(
-        &source_index_policy.policy_version,
-        source_index_policy.byte_cap,
-        source_index_policy.structural_unit_cap,
-    );
-    let source_candidates = source_exclusions
-        .iter()
-        .map(|record| {
-            let mut candidate = source_policy_exclusion_candidate(record);
-            candidate.policy_version = source_index_policy.policy_version.clone();
-            candidate.byte_cap = source_index_policy.byte_cap;
-            candidate.structural_unit_cap = source_index_policy.structural_unit_cap;
-            candidate
-        })
-        .collect::<Vec<_>>();
-    staged
-        .store_mut()
-        .publish_source_policy_exclusion_generation(
-            publication,
-            &selected_identity.project_id,
-            &selected_identity.workspace_id,
-            current_source_policy,
-            &source_candidates,
-        )
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to rebind pinned source-policy exclusions: {error}"
-            ))
-        })?;
-    staged
-        .store_mut()
-        .publish_structural_text_unit_generation(publication)
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to rebind pinned structural publication: {error}"
-            ))
-        })?;
-    staged
-        .store_mut()
-        .put_index_publication(publication)
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to persist staged semantic projection identity: {error}"
-            ))
-        })?;
-    Ok(dense_manifest.anchor_count)
-}
-
-fn commit_semantic_projection(
-    mut staged: StagedSnapshot,
-    storage_path: &Path,
-    expected_publication: &IndexPublicationRecord,
-    publication: &IndexPublicationRecord,
-    prepared_search_state: SearchStateBuildResult,
-    cancel_token: Option<&CancellationToken>,
-) -> Result<(SearchStateBuildResult, StagedSnapshotPublishStats, Duration), ApiError> {
-    let staged_path = staged.path().to_path_buf();
-    let result = (|| {
-        #[cfg(test)]
-        publication_test_checkpoint(PublicationTestBoundary::CatalogLock, cancel_token)?;
-        ensure_indexing_active(cancel_token)?;
-        #[cfg(test)]
-        run_semantic_projection_before_revalidate_hook(storage_path);
-        let _catalog_guard = SearchGenerationCatalogGuard::acquire(storage_path)?;
-        let live_publication =
-            Store::database_complete_index_publication(storage_path).map_err(|error| {
-                ApiError::internal(format!(
-                    "Failed to revalidate the pinned core before promotion: {error}"
-                ))
-            })?;
-        if live_publication.as_ref() != Some(expected_publication) {
-            return Err(ApiError::new(
-                "publication_changed",
-                "The live core changed while semantic projections were being rebuilt.",
-            ));
-        }
-        ensure_indexing_active(cancel_token)?;
-        #[cfg(test)]
-        publication_test_checkpoint(PublicationTestBoundary::MarkerCompletion, cancel_token)?;
-        ensure_indexing_active(cancel_token)?;
-        staged
-            .store_mut()
-            .finish_incremental_run()
-            .map_err(|error| {
-                ApiError::internal(format!(
-                    "Failed to complete the staged semantic projection marker: {error}"
-                ))
-            })?;
-        #[cfg(test)]
-        publication_test_checkpoint(PublicationTestBoundary::DatabaseReplacement, cancel_token)?;
-        ensure_indexing_active(cancel_token)?;
-        let publish_started = Instant::now();
-        let publish_stats = staged.publish_with_stats(storage_path).map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to publish staged semantic projections: {error}. Preserved staged snapshot at {}",
-                staged_path.display()
-            ))
-        })?;
-        Ok((
-            prepared_search_state,
-            publish_stats,
-            publish_started.elapsed(),
-        ))
-    })();
-    if result.is_err() {
-        discard_unpublished_search_generation(storage_path, publication);
-    }
-    result
-}
-
-fn semantic_projection_phase_timings(
-    prepared: &PreparedSemanticProjection,
-    publish_stats: StagedSnapshotPublishStats,
-    publish_duration: Duration,
-) -> IndexingPhaseTimings {
-    let mut phase_timings = IndexingPhaseTimings {
-        deferred_indexes_ms: Some(
-            prepared
-                .finalize_stats
-                .deferred_indexes_ms
-                .saturating_add(prepared.stats.semantic_context_index_ms),
-        ),
-        summary_snapshot_ms: Some(prepared.finalize_stats.summary_snapshot_ms),
-        detail_snapshot_ms: Some(prepared.detail_snapshot_ms),
-        publish_ms: Some(clamp_u128_to_u32(publish_duration.as_millis())),
-        staged_sqlite_wal_autocheckpoint_bytes: publish_stats.sqlite_wal_autocheckpoint_bytes,
-        staged_sqlite_checkpoint_ms: publish_stats.sqlite_checkpoint_ms,
-        staged_sqlite_sync_ms: publish_stats.sqlite_sync_ms,
-        staged_snapshot_copy: publish_stats
-            .snapshot_copy
-            .map(database_snapshot_copy_timings),
-        core_promotion: Some(core_promotion_timings(publish_stats.core_promotion)),
-        ..Default::default()
-    };
-    apply_semantic_projection_stats(&mut phase_timings, prepared.stats);
-    phase_timings
-}
-
-fn semantic_projection_republish_for_runtime(
-    root: &Path,
-    storage_path: &Path,
-    cancel_token: Option<&CancellationToken>,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-    source_index_policy: &SourceIndexPolicy,
-) -> Result<
-    (
-        IndexingRunSummary,
-        IndexPublicationRecord,
-        IndexPublicationRecord,
-        u32,
-        u64,
-    ),
-    ApiError,
-> {
-    ensure_indexing_active(cancel_token)?;
-    if !storage_path.is_file() {
-        return Err(ApiError::new(
-            "semantic_projection_core_missing",
-            "Semantic projection republish requires an existing complete core publication.",
-        ));
-    }
-
-    let expected_schema_version =
-        Store::database_schema_version(storage_path).map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to pin the stored core schema version: {error}"
-            ))
-        })?;
-
-    let expected_publication = Store::database_complete_index_publication(storage_path)
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to pin the complete core publication: {error}"
-            ))
-        })?
-        .ok_or_else(|| {
-            ApiError::new(
-                "semantic_projection_core_incomplete",
-                "Semantic projection republish requires a complete core publication.",
-            )
-        })?;
-    let mut staged = SnapshotStore::clone_live_to_staged(storage_path).map_err(|error| {
-        ApiError::internal(format!(
-            "Failed to clone the pinned core for semantic projection republish: {error}"
-        ))
-    })?;
-    let cleanup_staged_path = staged.path().to_path_buf();
-    let result = (|| {
-        let source_exclusions = validate_semantic_projection_core(
-            &mut staged,
-            root,
-            expected_schema_version,
-            &expected_publication,
-            source_index_policy,
-        )?;
-        let publication = next_index_publication(
-            Some(&expected_publication),
-            IndexPublicationMode::SemanticProjection,
-            &Uuid::new_v4().to_string(),
-        )?;
-        let prepared =
-            prepare_semantic_projection(&mut staged, &publication, cancel_token, runtime)?;
-        let dense_anchor_count = stage_semantic_projection_publication(
-            &mut staged,
-            root,
-            &publication,
-            &source_exclusions,
-            source_index_policy,
-            cancel_token,
-        )?;
-        let prepared_search_state = match rebuild_search_state_from_storage_for_runtime(
-            staged.store_mut(),
-            storage_path,
-            None,
-            false,
-            runtime,
-            cancel_token,
-            None,
-        ) {
-            Ok(state) => state,
-            Err(error) => {
-                discard_unpublished_search_generation(storage_path, &publication);
-                return Err(error);
-            }
-        };
-        let (prepared_search_state, publish_stats, publish_duration) = commit_semantic_projection(
-            staged,
-            storage_path,
-            &expected_publication,
-            &publication,
-            prepared_search_state,
-            cancel_token,
-        )?;
-        let phase_timings =
-            semantic_projection_phase_timings(&prepared, publish_stats, publish_duration);
-        Ok((
-            IndexingRunSummary {
-                phase_timings,
-                staged_semantic_stats: prepared.stats,
-                llm_refresh_scope: None,
-                #[cfg(test)]
-                publication: publication.clone(),
-                prepared_search_state: Some(prepared_search_state),
-            },
-            publication,
-            prepared.stats.symbol_search_docs_written,
-            dense_anchor_count,
-        ))
-    })();
-
-    match result {
-        Ok((summary, publication, symbol_document_count, dense_anchor_count)) => Ok((
-            summary,
-            expected_publication,
-            publication,
-            symbol_document_count,
-            dense_anchor_count,
-        )),
-        Err(error) => {
-            let _ = SnapshotStore::discard_staged(&cleanup_staged_path);
-            Err(error)
-        }
-    }
-}
-
-struct FullIndexLiveState {
-    previous_publication: Option<IndexPublicationRecord>,
-    publication: IndexPublicationRecord,
-    dense_anchor_source_identity: String,
-    recovering_incomplete_run: bool,
-    has_verified_publication: bool,
-}
-
-fn incomplete_live_index_requires_recovery(storage_path: &Path) -> Result<bool, ApiError> {
-    if !storage_path.exists() {
-        return Ok(false);
-    }
-    match Storage::database_schema_version(storage_path) {
-        Ok(version) if version > codestory_store::CURRENT_SCHEMA_VERSION => {
-            Storage::database_has_incomplete_incremental_run(storage_path).map_err(|error| {
-                ApiError::internal(format!("Failed to inspect live storage: {error}"))
-            })
-        }
-        Ok(_) => match Storage::database_has_incomplete_incremental_run(storage_path) {
-            Ok(marked) => Ok(marked),
-            Err(error) => {
-                tracing::warn!(
-                    path = %storage_path.display(),
-                    "Live storage could not be inspected; rebuilding without copying derived state: {error}"
-                );
-                Ok(true)
-            }
-        },
-        Err(error) => {
-            tracing::warn!(
-                path = %storage_path.display(),
-                "Live storage schema could not be read; rebuilding without copying derived state: {error}"
-            );
-            Ok(true)
-        }
-    }
-}
-
-fn live_publication_is_verified(
-    root: &Path,
-    storage_path: &Path,
-    expected: Option<&IndexPublicationRecord>,
-    recovering_incomplete_run: bool,
-    source_index_policy: &SourceIndexPolicy,
-) -> bool {
-    if recovering_incomplete_run {
-        return false;
-    }
-    let Some(expected) = expected else {
-        return false;
-    };
-    let live = match Store::open_read_only(storage_path) {
-        Ok(storage) => storage,
-        Err(error) => {
-            tracing::debug!(
-                path = %storage_path.display(),
-                "Live publication could not be opened for verification: {error}"
-            );
-            return false;
-        }
-    };
-    let publication = match live.get_complete_index_publication() {
-        Ok(Some(publication)) if publication == *expected => publication,
-        Ok(_) => return false,
-        Err(error) => {
-            tracing::debug!(
-                path = %storage_path.display(),
-                "Live core publication could not be verified: {error}"
-            );
-            return false;
-        }
-    };
-    if let Err(error) = live.validate_dense_anchor_publication(&publication) {
-        tracing::debug!(
-            path = %storage_path.display(),
-            "Live dense anchor publication could not be verified: {error}"
-        );
-        return false;
-    }
-    validate_structural_text_units(&live, &publication).is_ok()
-        && validate_source_policy_exclusions(&live, root, &publication, source_index_policy).is_ok()
-}
-
-fn inspect_full_index_live_state(
-    root: &Path,
-    storage_path: &Path,
-    source_index_policy: &SourceIndexPolicy,
-) -> Result<FullIndexLiveState, ApiError> {
-    let previous_publication = if storage_path.exists() {
-        Store::database_index_publication(storage_path).map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to inspect live publication identity: {error}"
-            ))
-        })?
-    } else {
-        None
-    };
-    let publication = next_index_publication(
-        previous_publication.as_ref(),
-        IndexPublicationMode::Full,
-        &Uuid::new_v4().to_string(),
-    )?;
-    let dense_anchor_source_identity =
-        format!("core:{}:{}", publication.generation_id, publication.run_id);
-    let recovering_incomplete_run = incomplete_live_index_requires_recovery(storage_path)?;
-    let has_verified_publication = live_publication_is_verified(
-        root,
-        storage_path,
-        previous_publication.as_ref(),
-        recovering_incomplete_run,
-        source_index_policy,
-    );
-    Ok(FullIndexLiveState {
-        previous_publication,
-        publication,
-        dense_anchor_source_identity,
-        recovering_incomplete_run,
-        has_verified_publication,
-    })
-}
-
-fn validate_full_refresh_coverage(
-    root: &Path,
-    staged: &mut StagedSnapshot,
-    live_state: &FullIndexLiveState,
-) -> Result<(), ApiError> {
-    let blocking_gaps = stored_file_coverage_diagnostics(root, staged.store_mut())?
-        .into_iter()
-        .filter(|entry| entry.reason != FileCoverageReason::ParserPartial)
-        .collect::<Vec<_>>();
-    if blocking_gaps.is_empty() {
-        return Ok(());
-    }
-    let sample = blocking_gaps
-        .iter()
-        .take(3)
-        .map(|entry| format!("{} ({})", entry.path, entry.reason.as_str()))
-        .collect::<Vec<_>>()
-        .join(", ");
-    let remainder = blocking_gaps.len().saturating_sub(3);
-    let sample = if remainder > 0 {
-        format!("{sample}, and {remainder} more")
-    } else {
-        sample
-    };
-    let preserved_state = if live_state.has_verified_publication {
-        "The previous complete publication was preserved"
-    } else if live_state.recovering_incomplete_run {
-        "The existing live index and its incomplete-run recovery fence were preserved"
-    } else if live_state.previous_publication.is_some() {
-        "The existing live index was preserved and no replacement publication was created"
-    } else {
-        "No core publication was created"
-    };
-    let count = blocking_gaps.len();
-    Err(ApiError::source_coverage_failure(
-        source_coverage_failure_code(&blocking_gaps),
-        format!(
-            "Effective refresh mode `full` could not verify {count} scheduled file(s): {sample}. {preserved_state}."
-        ),
-        blocking_gaps,
-    ))
-}
-
-fn copy_forward_full_refresh_artifacts(staged: &mut StagedSnapshot, storage_path: &Path) {
-    match staged
-        .store_mut()
-        .copy_retrieval_artifact_nodes_from(storage_path)
-    {
-        Ok(copied) => tracing::debug!(
-            copied,
-            "Copied retrieval artifact nodes into staged storage"
-        ),
-        Err(error) => {
-            tracing::warn!("Failed to copy retrieval artifact nodes into staged storage: {error}")
-        }
-    }
-    match staged
-        .store_mut()
-        .copy_symbol_search_docs_from(storage_path)
-    {
-        Ok(copied) => tracing::debug!(copied, "Copied symbol docs into staged storage"),
-        Err(error) => tracing::warn!("Failed to copy symbol docs into staged storage: {error}"),
-    }
-    match staged
-        .store_mut()
-        .copy_dense_anchor_inputs_from(storage_path)
-    {
-        Ok(copied) => tracing::debug!(copied, "Copied dense anchor inputs into staged storage"),
-        Err(error) => {
-            tracing::warn!("Failed to copy dense anchor inputs into staged storage: {error}")
-        }
-    }
-}
-
-struct PreparedFullRefreshSnapshots {
-    semantic_stats: SemanticProjectionStats,
-    finalize_stats: StagedSnapshotFinalizeStats,
-    detail_snapshot_ms: u32,
-    semantic_duration: Duration,
-    snapshot_duration: Duration,
-}
-
-fn prepare_full_refresh_snapshots(
-    staged: &mut StagedSnapshot,
-    source_identity: &str,
-    cancel_token: Option<&CancellationToken>,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-) -> Result<PreparedFullRefreshSnapshots, ApiError> {
-    let semantic_started = Instant::now();
-    let semantic_stats = finalize_staged_semantic_docs_for_runtime(
-        staged.store_mut(),
-        None,
-        None,
-        source_identity,
-        cancel_token,
-        runtime,
-        SemanticProjectionDocumentSource::SourceFiles,
-    )?;
-    ensure_indexing_active(cancel_token)?;
-    let semantic_duration = semantic_started.elapsed();
-    let snapshot_started = Instant::now();
-    let finalize_stats = staged.snapshots().finalize_staged().map_err(|error| {
-        ApiError::internal(format!(
-            "Failed to finalize staged snapshot lifecycle: {error}"
-        ))
-    })?;
-    let detail_started = Instant::now();
-    staged.snapshots().refresh_detail().map_err(|error| {
-        ApiError::internal(format!(
-            "Failed to finalize staged detail snapshots: {error}"
-        ))
-    })?;
-    ensure_indexing_active(cancel_token)?;
-    Ok(PreparedFullRefreshSnapshots {
-        semantic_stats,
-        finalize_stats,
-        detail_snapshot_ms: clamp_u128_to_u32(detail_started.elapsed().as_millis()),
-        semantic_duration,
-        snapshot_duration: snapshot_started.elapsed(),
-    })
-}
-
-fn stage_core_publication_identity(
-    staged: &mut StagedSnapshot,
-    root: &Path,
-    workspace: &WorkspaceManifest,
-    publication: &IndexPublicationRecord,
-    policy_exclusions: &[OversizedSourceExclusionCandidate],
-    source_index_policy: &SourceIndexPolicy,
-    cancel_token: Option<&CancellationToken>,
-) -> Result<(), ApiError> {
-    ensure_indexing_active(cancel_token)?;
-    #[cfg(test)]
-    publication_test_checkpoint(PublicationTestBoundary::Identity, cancel_token)?;
-    staged
-        .store_mut()
-        .publish_dense_anchor_generation(publication, SEMANTIC_POLICY_VERSION)
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to publish complete dense anchor inputs: {error}"
-            ))
-        })?;
-    #[cfg(test)]
-    run_source_policy_before_revalidate_hook();
-    let exclusions =
-        revalidate_source_policy_exclusions(workspace, policy_exclusions, source_index_policy)?;
-    publish_source_policy_exclusions(
-        staged.store_mut(),
-        root,
-        publication,
-        &exclusions,
-        source_index_policy,
-    )?;
-    staged
-        .store_mut()
-        .publish_structural_text_unit_generation(publication)
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to publish complete structural text units: {error}"
-            ))
-        })?;
-    let mode = match publication.mode {
-        IndexPublicationMode::Full => "full",
-        IndexPublicationMode::Incremental => "incremental",
-        IndexPublicationMode::SemanticProjection => "semantic projection",
-    };
-    staged
-        .store_mut()
-        .put_index_publication(publication)
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to persist staged {mode} publication identity: {error}"
-            ))
-        })
-}
-
-#[derive(Clone, Copy)]
-enum CoreCommitMode {
-    Full { finish_recovery_marker: bool },
-    Incremental,
-}
-
-struct PreparedCoreCommit {
-    staged: Option<StagedSnapshot>,
-    search_state: Option<SearchStateBuildResult>,
-    storage_path: PathBuf,
-    publication: IndexPublicationRecord,
-    committed: bool,
-}
-
-impl PreparedCoreCommit {
-    fn new(
-        staged: StagedSnapshot,
-        search_state: SearchStateBuildResult,
-        storage_path: &Path,
-        publication: &IndexPublicationRecord,
-    ) -> Self {
-        Self {
-            staged: Some(staged),
-            search_state: Some(search_state),
-            storage_path: storage_path.to_path_buf(),
-            publication: publication.clone(),
-            committed: false,
-        }
-    }
-
-    fn staged_mut(&mut self) -> &mut StagedSnapshot {
-        self.staged
-            .as_mut()
-            .expect("prepared core commit must own staged storage")
-    }
-
-    fn commit(
-        mut self,
-        mode: CoreCommitMode,
-        cancel_token: Option<&CancellationToken>,
-    ) -> Result<(SearchStateBuildResult, StagedSnapshotPublishStats, Duration), ApiError> {
-        #[cfg(test)]
-        publication_test_checkpoint(PublicationTestBoundary::CatalogLock, cancel_token)?;
-        let _catalog_guard = SearchGenerationCatalogGuard::acquire(&self.storage_path)?;
-        ensure_indexing_active(cancel_token)?;
-        let finish_marker = match mode {
-            CoreCommitMode::Full {
-                finish_recovery_marker,
-            } => finish_recovery_marker,
-            CoreCommitMode::Incremental => true,
-        };
-        if finish_marker {
-            #[cfg(test)]
-            publication_test_checkpoint(PublicationTestBoundary::MarkerCompletion, cancel_token)?;
-            ensure_indexing_active(cancel_token)?;
-            let marker = match mode {
-                CoreCommitMode::Full { .. } => "full-recovery",
-                CoreCommitMode::Incremental => "incremental",
-            };
-            self.staged_mut()
-                .store_mut()
-                .finish_incremental_run()
-                .map_err(|error| {
-                    ApiError::internal(format!(
-                        "Failed to complete staged {marker} marker: {error}"
-                    ))
-                })?;
-        }
-        #[cfg(test)]
-        publication_test_checkpoint(PublicationTestBoundary::DatabaseReplacement, cancel_token)?;
-        ensure_indexing_active(cancel_token)?;
-        let staged_path = self.staged_mut().path().to_path_buf();
-        let staged = self
-            .staged
-            .take()
-            .expect("prepared core commit must own staged storage");
-        let publish_started = Instant::now();
-        let publish_stats = staged
-            .publish_with_stats(&self.storage_path)
-            .map_err(|error| {
-                let publication = match mode {
-                    CoreCommitMode::Full { .. } => "storage",
-                    CoreCommitMode::Incremental => "incremental storage",
-                };
-                ApiError::internal(format!(
-                    "Failed to publish staged {publication}: {error}. Preserved staged snapshot at {}",
-                    staged_path.display()
-                ))
-            })?;
-        let search_state = self
-            .search_state
-            .take()
-            .expect("prepared core commit must own search state");
-        self.committed = true;
-        Ok((search_state, publish_stats, publish_started.elapsed()))
-    }
-}
-
-impl Drop for PreparedCoreCommit {
-    fn drop(&mut self) {
-        if self.committed {
-            return;
-        }
-        drop(self.search_state.take());
-        if let Some(staged) = self.staged.take() {
-            let _ = staged.discard();
-        }
-        discard_unpublished_search_generation(&self.storage_path, &self.publication);
-    }
-}
-
-fn apply_core_indexing_stats(timings: &mut IndexingPhaseTimings, stats: &IncrementalIndexingStats) {
-    timings.parse_index_ms = clamp_u64_to_u32(stats.parse_index_ms);
-    timings.projection_flush_ms = clamp_u64_to_u32(stats.projection_flush_ms);
-    timings.edge_resolution_ms = clamp_u64_to_u32(stats.edge_resolution_ms);
-    timings.error_flush_ms = clamp_u64_to_u32(stats.error_flush_ms);
-    timings.cleanup_ms = clamp_u64_to_u32(stats.cleanup_ms);
-    timings.artifact_cache_write_ms = Some(clamp_u64_to_u32(stats.artifact_cache_write_ms));
-    timings.artifact_cache_writes = Some(clamp_usize_to_u32(stats.artifact_cache_writes));
-    timings.artifact_cache_write_transactions =
-        Some(clamp_usize_to_u32(stats.artifact_cache_write_transactions));
-    timings.parser_artifact_cache =
-        Some(artifact_cache_access_timings(&stats.parser_artifact_cache));
-    timings.structural_artifact_cache = Some(artifact_cache_access_timings(
-        &stats.structural_artifact_cache,
-    ));
-    timings.source_prepare_ms = Some(clamp_u64_to_u32(stats.source_prepare_ms));
-    timings.projection_batch_wall_ms = Some(clamp_u64_to_u32(stats.projection_batch_wall_ms));
-    timings.projection_batch_transactions =
-        Some(clamp_usize_to_u32(stats.projection_batch_transactions));
-    timings.projection_persistence = Some(projection_persistence_timings(
-        &stats.projection_persistence,
-    ));
-    timings.unresolved_calls_start = clamp_usize_to_u32(stats.unresolved_calls_start);
-    timings.unresolved_imports_start = clamp_usize_to_u32(stats.unresolved_imports_start);
-    timings.resolved_calls = clamp_usize_to_u32(stats.resolved_calls);
-    timings.resolved_imports = clamp_usize_to_u32(stats.resolved_imports);
-    timings.unresolved_calls_end = clamp_usize_to_u32(stats.unresolved_calls_end);
-    timings.unresolved_imports_end = clamp_usize_to_u32(stats.unresolved_imports_end);
-}
-
-fn apply_resolution_telemetry(
-    timings: &mut IndexingPhaseTimings,
-    telemetry: &OptionalResolutionTelemetry,
-) {
-    timings.setup_existing_projection_ids_ms = telemetry.setup_existing_projection_ids_ms;
-    timings.setup_seed_symbol_table_ms = telemetry.setup_seed_symbol_table_ms;
-    timings.flush_files_ms = telemetry.flush_files_ms;
-    timings.flush_nodes_ms = telemetry.flush_nodes_ms;
-    timings.flush_edges_ms = telemetry.flush_edges_ms;
-    timings.flush_occurrences_ms = telemetry.flush_occurrences_ms;
-    timings.flush_component_access_ms = telemetry.flush_component_access_ms;
-    timings.flush_callable_projection_ms = telemetry.flush_callable_projection_ms;
-    timings.resolution_override_count_ms = telemetry.resolution_override_count_ms;
-    timings.resolution_unresolved_counts_ms = telemetry.resolution_unresolved_counts_ms;
-    timings.resolution_calls_ms = telemetry.resolution_calls_ms;
-    timings.resolution_imports_ms = telemetry.resolution_imports_ms;
-    timings.resolution_cleanup_ms = telemetry.resolution_cleanup_ms;
-    timings.resolution_call_candidate_index_ms = telemetry.resolution_call_candidate_index_ms;
-    timings.resolution_import_candidate_index_ms = telemetry.resolution_import_candidate_index_ms;
-    timings.resolution_call_semantic_index_ms = telemetry.resolution_call_semantic_index_ms;
-    timings.resolution_import_semantic_index_ms = telemetry.resolution_import_semantic_index_ms;
-    timings.resolution_support_snapshot_limit_bytes =
-        telemetry.resolution_support_snapshot_limit_bytes;
-    timings.resolution_support_snapshot_stored = telemetry.resolution_support_snapshot_stored;
-    timings.resolution_support_snapshot_skipped_oversize =
-        telemetry.resolution_support_snapshot_skipped_oversize;
-    timings.resolution_call_semantic_candidates_ms =
-        telemetry.resolution_call_semantic_candidates_ms;
-    timings.resolution_import_semantic_candidates_ms =
-        telemetry.resolution_import_semantic_candidates_ms;
-    timings.resolution_call_semantic_requests = telemetry.resolution_call_semantic_requests;
-    timings.resolution_call_semantic_unique_requests =
-        telemetry.resolution_call_semantic_unique_requests;
-    timings.resolution_call_semantic_skipped_requests =
-        telemetry.resolution_call_semantic_skipped_requests;
-    timings.resolution_import_semantic_requests = telemetry.resolution_import_semantic_requests;
-    timings.resolution_import_semantic_unique_requests =
-        telemetry.resolution_import_semantic_unique_requests;
-    timings.resolution_import_semantic_skipped_requests =
-        telemetry.resolution_import_semantic_skipped_requests;
-    timings.resolution_call_compute_ms = telemetry.resolution_call_compute_ms;
-    timings.resolution_import_compute_ms = telemetry.resolution_import_compute_ms;
-    timings.resolution_call_apply_ms = telemetry.resolution_call_apply_ms;
-    timings.resolution_import_apply_ms = telemetry.resolution_import_apply_ms;
-    timings.resolution_override_resolution_ms = telemetry.resolution_override_resolution_ms;
-    timings.resolved_calls_same_file = telemetry.resolved_calls_same_file;
-    timings.resolved_calls_same_module = telemetry.resolved_calls_same_module;
-    timings.resolved_calls_global_unique = telemetry.resolved_calls_global_unique;
-    timings.resolved_calls_semantic = telemetry.resolved_calls_semantic;
-    timings.resolved_imports_same_file = telemetry.resolved_imports_same_file;
-    timings.resolved_imports_same_module = telemetry.resolved_imports_same_module;
-    timings.resolved_imports_global_unique = telemetry.resolved_imports_global_unique;
-    timings.resolved_imports_fuzzy = telemetry.resolved_imports_fuzzy;
-    timings.resolved_imports_semantic = telemetry.resolved_imports_semantic;
-}
-
-fn apply_staged_publication_timings(
-    timings: &mut IndexingPhaseTimings,
-    finalize_stats: StagedSnapshotFinalizeStats,
-    detail_snapshot_ms: u32,
-    publish_stats: StagedSnapshotPublishStats,
-    publish_duration: Duration,
-    semantic_context_index_ms: u32,
-) {
-    timings.deferred_indexes_ms = Some(
-        finalize_stats
-            .deferred_indexes_ms
-            .saturating_add(semantic_context_index_ms),
-    );
-    timings.summary_snapshot_ms = Some(finalize_stats.summary_snapshot_ms);
-    timings.detail_snapshot_ms = Some(detail_snapshot_ms);
-    timings.publish_ms = Some(clamp_u128_to_u32(publish_duration.as_millis()));
-    timings.staged_sqlite_wal_autocheckpoint_bytes = publish_stats.sqlite_wal_autocheckpoint_bytes;
-    timings.staged_sqlite_checkpoint_ms = publish_stats.sqlite_checkpoint_ms;
-    timings.staged_sqlite_sync_ms = publish_stats.sqlite_sync_ms;
-    timings.staged_snapshot_copy = publish_stats
-        .snapshot_copy
-        .map(database_snapshot_copy_timings);
-    timings.core_promotion = Some(core_promotion_timings(publish_stats.core_promotion));
-}
-
-fn apply_full_refresh_pipeline_timings(
-    timings: &mut IndexingPhaseTimings,
-    stats: &IncrementalIndexingStats,
-    wall: FullRefreshWallTimings,
-) {
-    timings.full_refresh_wall = Some(wall);
-    let pipeline_enabled = stats.full_refresh_queue_capacity > 0;
-    timings.full_refresh_chunks_produced =
-        pipeline_enabled.then_some(clamp_usize_to_u32(stats.full_refresh_chunks_produced));
-    timings.full_refresh_chunks_persisted =
-        pipeline_enabled.then_some(clamp_usize_to_u32(stats.full_refresh_chunks_persisted));
-    timings.full_refresh_queue_capacity =
-        pipeline_enabled.then_some(clamp_usize_to_u32(stats.full_refresh_queue_capacity));
-    timings.full_refresh_queue_high_water =
-        pipeline_enabled.then_some(clamp_usize_to_u32(stats.full_refresh_queue_high_water));
-    timings.full_refresh_producer_blocked_ms =
-        pipeline_enabled.then_some(clamp_u64_to_u32(stats.full_refresh_producer_blocked_ms));
-    timings.full_refresh_writer_idle_ms =
-        pipeline_enabled.then_some(clamp_u64_to_u32(stats.full_refresh_writer_idle_ms));
-    let chunking_enabled = stats.full_refresh_chunk_target_bytes > 0;
-    timings.full_refresh_chunk_target_bytes =
-        chunking_enabled.then_some(stats.full_refresh_chunk_target_bytes);
-    timings.full_refresh_chunk_target_nodes =
-        chunking_enabled.then_some(clamp_usize_to_u32(stats.full_refresh_chunk_target_nodes));
-    timings.full_refresh_chunk_file_ceiling =
-        chunking_enabled.then_some(clamp_usize_to_u32(stats.full_refresh_chunk_file_ceiling));
-    timings.full_refresh_chunk_max_files =
-        chunking_enabled.then_some(clamp_usize_to_u32(stats.full_refresh_chunk_max_files));
-    timings.full_refresh_chunk_max_planned_bytes =
-        chunking_enabled.then_some(stats.full_refresh_chunk_max_planned_bytes);
-    timings.full_refresh_chunk_max_nodes =
-        chunking_enabled.then_some(clamp_usize_to_u32(stats.full_refresh_chunk_max_nodes));
-    timings.full_refresh_chunk_budget_overruns =
-        chunking_enabled.then_some(clamp_usize_to_u32(stats.full_refresh_chunk_budget_overruns));
-    timings.full_refresh_chunk_planning_ms =
-        chunking_enabled.then_some(clamp_u64_to_u32(stats.full_refresh_chunk_planning_ms));
-}
-
-fn core_indexing_phase_timings(
-    stats: &IncrementalIndexingStats,
-    finalize_stats: StagedSnapshotFinalizeStats,
-    detail_snapshot_ms: u32,
-    publish_stats: StagedSnapshotPublishStats,
-    publish_duration: Duration,
-    semantic_context_index_ms: u32,
-) -> IndexingPhaseTimings {
-    let mut timings = IndexingPhaseTimings::default();
-    apply_core_indexing_stats(&mut timings, stats);
-    apply_resolution_telemetry(
-        &mut timings,
-        &OptionalResolutionTelemetry::from_incremental_stats(stats),
-    );
-    apply_staged_publication_timings(
-        &mut timings,
-        finalize_stats,
-        detail_snapshot_ms,
-        publish_stats,
-        publish_duration,
-        semantic_context_index_ms,
-    );
-    timings
-}
-
-struct StagedPreparation {
-    staged: Option<StagedSnapshot>,
-}
-
-impl StagedPreparation {
-    fn new(staged: StagedSnapshot) -> Self {
-        Self {
-            staged: Some(staged),
-        }
-    }
-
-    fn staged_mut(&mut self) -> &mut StagedSnapshot {
-        self.staged
-            .as_mut()
-            .expect("staged preparation must own staged storage")
-    }
-
-    fn release(mut self) -> StagedSnapshot {
-        self.staged
-            .take()
-            .expect("staged preparation must own staged storage")
-    }
-}
-
-impl Drop for StagedPreparation {
-    fn drop(&mut self) {
-        if let Some(staged) = self.staged.take() {
-            let _ = staged.discard();
-        }
-    }
-}
-
-struct FullRefreshIndexingOutput {
-    staged: StagedSnapshot,
-    stats: IncrementalIndexingStats,
-    policy_exclusions: Vec<OversizedSourceExclusionCandidate>,
-}
-
-fn run_full_refresh_indexer(
-    root: &Path,
-    storage_path: &Path,
-    events_tx: &Sender<AppEventPayload>,
-    cancel_token: Option<&CancellationToken>,
-    source_index_policy: &SourceIndexPolicy,
-    execution_plan: &RefreshExecutionPlan,
-    mut policy_exclusions: Vec<OversizedSourceExclusionCandidate>,
-    live_state: &FullIndexLiveState,
-    wall_durations: &mut FullRefreshWallDurations,
-) -> Result<FullRefreshIndexingOutput, ApiError> {
-    let stage_started = Instant::now();
-    let total_files = execution_plan.files_to_index.len().min(u32::MAX as usize) as u32;
-    let _ = events_tx.send(AppEventPayload::IndexingStarted {
-        file_count: total_files,
-    });
-    #[cfg(test)]
-    run_source_policy_after_plan_hook();
-    let staged = SnapshotStore::open_disposable_full_refresh(storage_path)
-        .map_err(|error| ApiError::internal(format!("Failed to open staged storage: {error}")))?;
-    let mut preparation = StagedPreparation::new(staged);
-    #[cfg(test)]
-    run_full_refresh_staged_store_hook(preparation.staged_mut().store_mut());
-    let copied_structural_artifacts = if live_state.has_verified_publication {
-        match preparation
-            .staged_mut()
-            .store_mut()
-            .copy_structural_text_artifact_cache_from(storage_path)
-        {
-            Ok(copied) => {
-                tracing::debug!(
-                    copied,
-                    "Copied verified structural artifacts into staged storage"
-                );
-                copied
-            }
-            Err(error) => {
-                tracing::warn!(
-                    "Failed to copy verified structural artifacts into staged storage; recollecting: {error}"
-                );
-                0
-            }
-        }
-    } else {
-        0
-    };
-    let bus = EventBus::new();
-    let forwarder = spawn_progress_forwarder(bus.receiver(), events_tx.clone());
-    let indexer = V2WorkspaceIndexer::new(root.to_path_buf())
-        .with_source_index_policy(source_index_policy.clone())
-        .with_artifact_cache_policies(ArtifactCachePolicies {
-            parser: ArtifactCachePolicy::KnownEmpty,
-            structural: if copied_structural_artifacts > 0 {
-                ArtifactCachePolicy::ReadThrough
-            } else {
-                ArtifactCachePolicy::KnownEmpty
-            },
-        });
-    wall_durations.stage_open = stage_started.elapsed();
-    let execution_started = Instant::now();
-    let result = indexer.run_with_policy_exclusions(
-        preparation.staged_mut().store_mut(),
-        execution_plan,
-        &bus,
-        cancel_token,
-    );
-    drop(bus);
-    let _ = forwarder.join();
-    let outcome = match result {
-        Ok(_) if is_indexing_cancelled(cancel_token) => {
-            return Err(indexing_cancelled_error());
-        }
-        Ok(outcome) => outcome,
-        Err(_) if is_indexing_cancelled(cancel_token) => {
-            return Err(indexing_cancelled_error());
-        }
-        Err(error) => return Err(ApiError::internal(format!("Indexing failed: {error}"))),
-    };
-    wall_durations.indexer_execution = execution_started.elapsed();
-    policy_exclusions.extend(outcome.policy_exclusions);
-    Ok(FullRefreshIndexingOutput {
-        staged: preparation.release(),
-        stats: outcome.stats,
-        policy_exclusions,
-    })
-}
-
-struct PreparedFullRefresh {
-    staged: StagedSnapshot,
-    live_state: FullIndexLiveState,
-    workspace: WorkspaceManifest,
-    stats: IncrementalIndexingStats,
-    policy_exclusions: Vec<OversizedSourceExclusionCandidate>,
-    snapshots: PreparedFullRefreshSnapshots,
-    wall_durations: FullRefreshWallDurations,
-    core_refresh_started: Instant,
-}
-
-fn prepare_full_refresh(
-    root: &Path,
-    storage_path: &Path,
-    events_tx: &Sender<AppEventPayload>,
-    cancel_token: Option<&CancellationToken>,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-    source_index_policy: &SourceIndexPolicy,
-) -> Result<PreparedFullRefresh, ApiError> {
-    let core_refresh_started = Instant::now();
-    let live_started = Instant::now();
-    let live_state = inspect_full_index_live_state(root, storage_path, source_index_policy)?;
-    let mut wall_durations = FullRefreshWallDurations {
-        live_inspection: live_started.elapsed(),
-        ..Default::default()
-    };
-    let discovery_started = Instant::now();
-    let workspace = runtime_workspace_manifest(root, storage_path)
-        .map_err(|error| ApiError::internal(format!("Failed to open project: {error}")))?;
-    let (execution_plan, policy_exclusions) =
-        full_refresh_execution_plan_with_coverage(root, &workspace, source_index_policy)?;
-    wall_durations.source_discovery = discovery_started.elapsed();
-    let output = run_full_refresh_indexer(
-        root,
-        storage_path,
-        events_tx,
-        cancel_token,
-        source_index_policy,
-        &execution_plan,
-        policy_exclusions,
-        &live_state,
-        &mut wall_durations,
-    )?;
-    let mut preparation = StagedPreparation::new(output.staged);
-    let coverage_started = Instant::now();
-    validate_full_refresh_coverage(root, preparation.staged_mut(), &live_state)?;
-    wall_durations.coverage_validation = coverage_started.elapsed();
-    let copy_started = Instant::now();
-    if !live_state.recovering_incomplete_run && storage_path.exists() {
-        copy_forward_full_refresh_artifacts(preparation.staged_mut(), storage_path);
-    }
-    wall_durations.copy_forward = copy_started.elapsed();
-    let snapshots = prepare_full_refresh_snapshots(
-        preparation.staged_mut(),
-        &live_state.dense_anchor_source_identity,
-        cancel_token,
-        runtime,
-    )?;
-    wall_durations.semantic_stage = snapshots.semantic_duration;
-    wall_durations.snapshot_stage = snapshots.snapshot_duration;
-    Ok(PreparedFullRefresh {
-        staged: preparation.release(),
-        live_state,
-        workspace,
-        stats: output.stats,
-        policy_exclusions: output.policy_exclusions,
-        snapshots,
-        wall_durations,
-        core_refresh_started,
-    })
-}
-
-fn index_full_for_runtime(
-    root: &Path,
-    storage_path: &Path,
-    events_tx: &Sender<AppEventPayload>,
-    cancel_token: Option<&CancellationToken>,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-    source_index_policy: &SourceIndexPolicy,
-) -> Result<IndexingRunSummary, ApiError> {
-    let PreparedFullRefresh {
-        mut staged,
-        live_state,
-        workspace,
-        stats: index_stats,
-        policy_exclusions,
-        snapshots: prepared_snapshots,
-        mut wall_durations,
-        core_refresh_started,
-    } = prepare_full_refresh(
-        root,
-        storage_path,
-        events_tx,
-        cancel_token,
-        runtime,
-        source_index_policy,
-    )?;
-    let publication = &live_state.publication;
-    let recovering_incomplete_run = live_state.recovering_incomplete_run;
-    let mut wall_stage_started = Instant::now();
-    if recovering_incomplete_run && let Err(err) = staged.store_mut().begin_incremental_run() {
-        let _ = staged.discard();
-        return Err(ApiError::internal(format!(
-            "Failed to preserve incomplete marker through staged recovery: {err}"
-        )));
-    }
-    if let Err(error) = stage_core_publication_identity(
-        &mut staged,
-        root,
-        &workspace,
-        publication,
-        &policy_exclusions,
-        source_index_policy,
-        cancel_token,
-    ) {
-        let _ = staged.discard();
-        return Err(error);
-    }
-    wall_durations.publication_prepare = wall_stage_started.elapsed();
-    wall_stage_started = Instant::now();
-    let prepared_search_state = match rebuild_search_state_from_storage_for_runtime(
-        staged.store_mut(),
-        storage_path,
-        None,
-        false,
-        runtime,
-        cancel_token,
-        None,
-    ) {
-        Ok(state) => state,
-        Err(error) => {
-            let _ = staged.discard();
-            discard_unpublished_search_generation(storage_path, &publication);
-            return Err(error);
-        }
-    };
-    if is_indexing_cancelled(cancel_token) {
-        drop(prepared_search_state);
-        let _ = staged.discard();
-        discard_unpublished_search_generation(storage_path, &publication);
-        return Err(indexing_cancelled_error());
-    }
-    wall_durations.search_generation = wall_stage_started.elapsed();
-    wall_stage_started = Instant::now();
-    let prepared_commit =
-        PreparedCoreCommit::new(staged, prepared_search_state, storage_path, publication);
-    let (prepared_search_state, staged_publish_stats, publish_duration) = prepared_commit.commit(
-        CoreCommitMode::Full {
-            finish_recovery_marker: recovering_incomplete_run,
-        },
-        cancel_token,
-    )?;
-    wall_durations.catalog_publication = wall_stage_started.elapsed();
-    let full_refresh_wall = wall_durations.finish(core_refresh_started.elapsed());
-    let mut phase_timings = core_indexing_phase_timings(
-        &index_stats,
-        prepared_snapshots.finalize_stats,
-        prepared_snapshots.detail_snapshot_ms,
-        staged_publish_stats,
-        publish_duration,
-        prepared_snapshots.semantic_stats.semantic_context_index_ms,
-    );
-    apply_full_refresh_pipeline_timings(&mut phase_timings, &index_stats, full_refresh_wall);
-    Ok(IndexingRunSummary {
-        phase_timings,
-        staged_semantic_stats: prepared_snapshots.semantic_stats,
-        llm_refresh_scope: None,
-        #[cfg(test)]
-        publication: publication.clone(),
-        prepared_search_state: Some(prepared_search_state),
-    })
-}
-
-#[cfg(test)]
-fn index_incremental(
-    root: &Path,
-    storage_path: &Path,
-    events_tx: &Sender<AppEventPayload>,
-    cancel_token: Option<&CancellationToken>,
-) -> Result<IndexingRunSummary, ApiError> {
-    index_incremental_for_runtime(
-        root,
-        storage_path,
-        events_tx,
-        cancel_token,
-        &test_sidecar_runtime_from_env(),
-        &SourceIndexPolicy::default(),
-    )
-}
-
-fn index_incremental_for_runtime(
-    root: &Path,
-    storage_path: &Path,
-    events_tx: &Sender<AppEventPayload>,
-    cancel_token: Option<&CancellationToken>,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-    source_index_policy: &SourceIndexPolicy,
-) -> Result<IndexingRunSummary, ApiError> {
-    run_incremental_indexing_common(
-        root,
-        storage_path,
-        events_tx,
-        cancel_token,
-        runtime,
-        source_index_policy,
-    )
-}
-
-fn spawn_progress_forwarder(
-    rx: Receiver<Event>,
-    progress_tx: Sender<AppEventPayload>,
-) -> std::thread::JoinHandle<()> {
-    std::thread::spawn(move || {
-        while let Ok(ev) = rx.recv() {
-            match ev {
-                Event::IndexingProgress { current, total } => {
-                    let _ = progress_tx.send(AppEventPayload::IndexingProgress {
-                        current: current.min(u32::MAX as usize) as u32,
-                        total: total.min(u32::MAX as usize) as u32,
-                    });
-                }
-                Event::StatusUpdate { message } => {
-                    let _ = progress_tx.send(AppEventPayload::StatusUpdate { message });
-                }
-                _ => {}
-            }
-        }
-    })
-}
-
-const FULL_REFRESH_REQUIRED_ERROR_CODE: &str = "full_refresh_required";
-
-fn full_refresh_required_error(
-    root: &Path,
-    reason_code: &str,
-    reason: impl AsRef<str>,
-) -> ApiError {
-    let project = root.to_string_lossy().to_string();
-    let next_command = format!(
-        "codestory-cli index --project {} --refresh full",
-        quote_refresh_command_argument(&project)
-    );
-    ApiError::with_details(
-        FULL_REFRESH_REQUIRED_ERROR_CODE,
-        format!(
-            "Refresh compatibility rejected the request before workspace reads: requested=incremental effective=none required=full reason={}",
-            reason.as_ref()
-        ),
-        ApiErrorDetails {
-            cause_code: Some(reason_code.to_string()),
-            failed_layer: Some("core_publication_compatibility".to_string()),
-            project: Some(project),
-            next_commands: vec![next_command.clone()],
-            minimum_next: vec![next_command.clone()],
-            full_repair: vec![next_command],
-            readiness: None,
-            embedding_capacity: None,
-            embedding_retry: None,
-            coverage_gaps: Vec::new(),
-        },
-    )
-}
-
-#[cfg(windows)]
-fn quote_refresh_command_argument(value: &str) -> String {
-    format!("'{}'", value.replace('\'', "''"))
-}
-
-#[cfg(not(windows))]
-fn quote_refresh_command_argument(value: &str) -> String {
-    format!("'{}'", value.replace('\'', "'\\''"))
-}
-
-fn ensure_incremental_refresh_compatible(root: &Path, storage_path: &Path) -> Result<(), ApiError> {
-    if !storage_path.is_file() {
-        return Err(full_refresh_required_error(
-            root,
-            "complete_core_publication_missing",
-            "complete_core_publication_missing",
-        ));
-    }
-    let schema_version = Store::database_schema_version_observational(storage_path).map_err(
-        |error| {
-            ApiError::internal(format!(
-                "Failed to inspect incremental refresh schema compatibility without recovery: {error}"
-            ))
-        },
-    )?;
-    if schema_version < CURRENT_SCHEMA_VERSION {
-        let (reason_code, reason) = if schema_version == 0 {
-            (
-                "complete_core_publication_missing",
-                "complete_core_publication_missing".to_string(),
-            )
-        } else {
-            (
-                "core_schema_upgrade_required",
-                format!(
-                    "core_schema_upgrade_required:observed={schema_version}:required={CURRENT_SCHEMA_VERSION}"
-                ),
-            )
-        };
-        return Err(full_refresh_required_error(root, reason_code, reason));
-    }
-    let storage = Store::open_freshness_observational(storage_path).map_err(|error| {
-        ApiError::internal(format!(
-            "Failed to inspect incremental refresh compatibility: {error}"
-        ))
-    })?;
-    if storage.has_incomplete_incremental_run().map_err(|error| {
-        ApiError::internal(format!(
-            "Failed to inspect incomplete incremental marker: {error}"
-        ))
-    })? {
-        return Err(full_refresh_required_error(
-            root,
-            "incomplete_incremental_publication",
-            "incomplete_incremental_publication",
-        ));
-    }
-    let Some(publication) = storage.get_complete_index_publication().map_err(|error| {
-        ApiError::internal(format!(
-            "Failed to inspect complete core publication: {error}"
-        ))
-    })?
-    else {
-        return Err(full_refresh_required_error(
-            root,
-            "complete_core_publication_missing",
-            "complete_core_publication_missing",
-        ));
-    };
-    if let Err(error) = storage.validate_structural_text_unit_publication(&publication) {
-        return Err(full_refresh_required_error(
-            root,
-            "structural_publication_incompatible",
-            format!("structural_publication_incompatible:{error}"),
-        ));
-    }
-    Ok(())
-}
-
-fn incremental_execution_plan(
-    staged: &mut StagedSnapshot,
-    root: &Path,
-    storage_path: &Path,
-    source_index_policy: &SourceIndexPolicy,
-) -> Result<(RefreshExecutionPlan, Vec<OversizedSourceExclusionCandidate>), ApiError> {
-    let workspace = runtime_workspace_manifest(root, storage_path)
-        .map_err(|error| ApiError::internal(format!("Failed to open project: {error}")))?;
-    let refresh_inputs = workspace_refresh_inputs(staged.store_mut())?;
-    let policy_refresh = workspace
-        .build_execution_outcome_with_policy(&refresh_inputs, source_index_policy)
-        .map_err(|error| ApiError::internal(format!("Failed to generate refresh info: {error}")))?;
-    if policy_refresh.refresh.inventory_outcome == WorkspaceInventoryOutcome::Complete {
-        return Ok((
-            policy_refresh.refresh.plan,
-            policy_refresh.policy_exclusions,
-        ));
-    }
-    let reason =
-        if policy_refresh.refresh.inventory_outcome == WorkspaceInventoryOutcome::Unreadable {
-            FileCoverageReason::Unreadable
-        } else {
-            FileCoverageReason::DiscoveryIncomplete
-        };
-    let mut gaps = policy_refresh
-        .refresh
-        .inventory_issues
-        .iter()
-        .map(|issue| FileCoverageDiagnosticDto {
-            path: runtime_relative_path(root, &issue.path),
-            reason,
-            retryable: file_coverage_retryable(reason),
-            verified_source: false,
-            projection_available: false,
-        })
-        .collect::<Vec<_>>();
-    if gaps.is_empty() {
-        gaps.push(FileCoverageDiagnosticDto {
-            path: ".".into(),
-            reason,
-            retryable: file_coverage_retryable(reason),
-            verified_source: false,
-            projection_available: false,
-        });
-    }
-    Err(ApiError::source_coverage_failure(
-        source_coverage_failure_code(&gaps),
-        format!(
-            "Incremental refresh requires a complete source inventory; discovery was {:?}.",
-            policy_refresh.refresh.inventory_outcome
-        ),
-        gaps,
-    ))
-}
-
-struct IncrementalSemanticPlan {
-    previous_indexed_file_ids_by_path: HashMap<String, codestory_contracts::graph::NodeId>,
-    policy_excluded_seed_file_ids: HashSet<codestory_contracts::graph::NodeId>,
-    previous_dependents_by_seed:
-        HashMap<codestory_contracts::graph::NodeId, HashSet<codestory_contracts::graph::NodeId>>,
-    component_reports: ComponentReportRefreshScope,
-}
-
-fn plan_incremental_semantics(
-    staged: &mut StagedSnapshot,
-    root: &Path,
-    execution_plan: &RefreshExecutionPlan,
-) -> Result<IncrementalSemanticPlan, ApiError> {
-    let mut planned_seed_file_ids = execution_plan
-        .files_to_remove
-        .iter()
-        .copied()
-        .map(codestory_contracts::graph::NodeId)
-        .collect::<HashSet<_>>();
-    let mut previous_indexed_file_ids_by_path = HashMap::new();
-    for path in &execution_plan.files_to_index {
-        let normalized_path = if path.is_absolute() {
-            path.clone()
-        } else {
-            root.join(path)
-        };
-        if let Some(file_info) = staged
-            .store_mut()
-            .get_file_by_path(&normalized_path)
-            .map_err(|error| {
-                ApiError::internal(format!(
-                    "Failed to resolve previous semantic scope for {}: {error}",
-                    normalized_path.display()
-                ))
-            })?
-        {
-            let file_id = codestory_contracts::graph::NodeId(file_info.id);
-            planned_seed_file_ids.insert(file_id);
-            previous_indexed_file_ids_by_path
-                .insert(runtime_relative_path(root, &normalized_path), file_id);
-        }
-    }
-    let previous_dependents_by_seed =
-        semantic_graph_dependent_file_ids_by_seed(staged.store_mut(), &planned_seed_file_ids)?;
-    let existing_file_paths = semantic_file_table_path_map(
-        staged
-            .store_mut()
-            .get_files()
-            .map_err(|error| ApiError::internal(format!("Failed to load files: {error}")))?,
-    );
-    let mut removed_component_keys = HashSet::new();
-    for file_id in &execution_plan.files_to_remove {
-        let path = existing_file_paths
-            .get(&codestory_contracts::graph::NodeId(*file_id))
-            .ok_or_else(|| {
-                ApiError::internal(format!(
-                    "Removed file is missing from staged component scope: {file_id}"
-                ))
-            })?;
-        if let Some(component_key) = semantic_component_key_for_path(Some(path)) {
-            removed_component_keys.insert(component_key);
-        }
-    }
-    Ok(IncrementalSemanticPlan {
-        previous_indexed_file_ids_by_path,
-        policy_excluded_seed_file_ids: HashSet::new(),
-        previous_dependents_by_seed,
-        component_reports: ComponentReportRefreshScope {
-            previous_file_paths: existing_file_paths,
-            removed_component_keys,
-        },
-    })
-}
-
-fn run_incremental_indexer(
-    staged: &mut StagedSnapshot,
-    root: &Path,
-    events_tx: &Sender<AppEventPayload>,
-    cancel_token: Option<&CancellationToken>,
-    source_index_policy: &SourceIndexPolicy,
-    execution_plan: &RefreshExecutionPlan,
-    semantic_plan: &mut IncrementalSemanticPlan,
-    policy_exclusions: &mut Vec<OversizedSourceExclusionCandidate>,
-) -> Result<IncrementalIndexingStats, ApiError> {
-    let total_files = execution_plan.files_to_index.len().min(u32::MAX as usize) as u32;
-    let _ = events_tx.send(AppEventPayload::IndexingStarted {
-        file_count: total_files,
-    });
-    #[cfg(test)]
-    run_incremental_staged_store_hook(staged.store_mut());
-    let bus = EventBus::new();
-    let forwarder = spawn_progress_forwarder(bus.receiver(), events_tx.clone());
-    if let Err(error) = ensure_indexing_active(cancel_token) {
-        drop(bus);
-        let _ = forwarder.join();
-        return Err(error);
-    }
-    let result = V2WorkspaceIndexer::new(root.to_path_buf())
-        .with_source_index_policy(source_index_policy.clone())
-        .run_with_policy_exclusions(staged.store_mut(), execution_plan, &bus, cancel_token);
-    drop(bus);
-    let _ = forwarder.join();
-    let outcome = match result {
-        Ok(_) if is_indexing_cancelled(cancel_token) => return Err(indexing_cancelled_error()),
-        Ok(outcome) => outcome,
-        Err(_) if is_indexing_cancelled(cancel_token) => return Err(indexing_cancelled_error()),
-        Err(error) => return Err(ApiError::internal(format!("Indexing failed: {error}"))),
-    };
-    for exclusion in &outcome.policy_exclusions {
-        if let Some(file_id) = semantic_plan
-            .previous_indexed_file_ids_by_path
-            .get(&exclusion.normalized_path)
-        {
-            semantic_plan.policy_excluded_seed_file_ids.insert(*file_id);
-            if let Some(component_key) = semantic_plan
-                .component_reports
-                .previous_file_paths
-                .get(file_id)
-                .and_then(|path| semantic_component_key_for_path(Some(path)))
-            {
-                semantic_plan
-                    .component_reports
-                    .removed_component_keys
-                    .insert(component_key);
-            }
-        }
-    }
-    policy_exclusions.extend(outcome.policy_exclusions);
-    Ok(outcome.stats)
-}
-
-fn validate_incremental_refresh_coverage(
-    staged: &mut StagedSnapshot,
-    root: &Path,
-) -> Result<(), ApiError> {
-    let blocking_gaps = stored_file_coverage_diagnostics(root, staged.store_mut())?
-        .into_iter()
-        .filter(|entry| entry.reason != FileCoverageReason::ParserPartial)
-        .collect::<Vec<_>>();
-    if blocking_gaps.is_empty() {
-        return Ok(());
-    }
-    let count = blocking_gaps.len();
-    let sample = blocking_gaps
-        .iter()
-        .take(3)
-        .map(|entry| format!("{} ({})", entry.path, entry.reason.as_str()))
-        .collect::<Vec<_>>()
-        .join(", ");
-    Err(ApiError::source_coverage_failure(
-        source_coverage_failure_code(&blocking_gaps),
-        format!(
-            "Incremental refresh could not verify {count} scheduled file(s): {sample}. The previous complete publication was preserved."
-        ),
-        blocking_gaps,
-    ))
-}
-
-fn incremental_semantic_refresh_scope(
-    staged: &mut StagedSnapshot,
-    root: &Path,
-    execution_plan: &RefreshExecutionPlan,
-    semantic_plan: &IncrementalSemanticPlan,
-) -> Result<HashSet<codestory_contracts::graph::NodeId>, ApiError> {
-    let mut refresh_seed_file_ids = HashSet::new();
-    for path in &execution_plan.files_to_index {
-        let normalized_path = if path.is_absolute() {
-            path.clone()
-        } else {
-            root.join(path)
-        };
-        let file_info = staged
-            .store_mut()
-            .get_file_by_path(&normalized_path)
-            .map_err(|error| {
-                ApiError::internal(format!(
-                    "Failed to resolve indexed semantic scope for {}: {error}",
-                    normalized_path.display()
-                ))
-            })?;
-        if let Some(file_info) = file_info
-            && file_info.complete
-        {
-            refresh_seed_file_ids.insert(codestory_contracts::graph::NodeId(file_info.id));
-        }
-    }
-    refresh_seed_file_ids.extend(
-        execution_plan
-            .files_to_remove
-            .iter()
-            .copied()
-            .map(codestory_contracts::graph::NodeId),
-    );
-    refresh_seed_file_ids.extend(&semantic_plan.policy_excluded_seed_file_ids);
-    let current_dependents_by_seed =
-        semantic_graph_dependent_file_ids_by_seed(staged.store_mut(), &refresh_seed_file_ids)?;
-    let mut refresh_scope = refresh_seed_file_ids.clone();
-    for seed_file_id in &refresh_seed_file_ids {
-        if let Some(file_ids) = semantic_plan.previous_dependents_by_seed.get(seed_file_id) {
-            refresh_scope.extend(file_ids);
-        }
-        if let Some(file_ids) = current_dependents_by_seed.get(seed_file_id) {
-            refresh_scope.extend(file_ids);
-        }
-    }
-    Ok(refresh_scope)
-}
-
-struct PreparedIncrementalRefresh {
-    staged: StagedSnapshot,
-    publication: IndexPublicationRecord,
-    stats: IncrementalIndexingStats,
-    finalize_stats: StagedSnapshotFinalizeStats,
-    detail_snapshot_ms: u32,
-    semantic_stats: SemanticProjectionStats,
-    semantic_refresh_scope: HashSet<codestory_contracts::graph::NodeId>,
-    policy_exclusions: Vec<OversizedSourceExclusionCandidate>,
-}
-
-fn prepare_incremental_refresh(
-    root: &Path,
-    storage_path: &Path,
-    events_tx: &Sender<AppEventPayload>,
-    cancel_token: Option<&CancellationToken>,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-    source_index_policy: &SourceIndexPolicy,
-) -> Result<PreparedIncrementalRefresh, ApiError> {
-    ensure_incremental_refresh_compatible(root, storage_path)?;
-    ensure_indexing_active(cancel_token)?;
-    let staged = SnapshotStore::clone_live_to_staged(storage_path).map_err(|error| {
-        ApiError::internal(format!(
-            "Failed to clone live storage for incremental build: {error}"
-        ))
-    })?;
-    let mut preparation = StagedPreparation::new(staged);
-    let previous_publication = preparation
-        .staged_mut()
-        .store_mut()
-        .get_index_publication()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to read staged publication identity: {error}"
-            ))
-        })?;
-    let rebuild_complete_dense_anchor_set = preparation
-        .staged_mut()
-        .store_mut()
-        .get_dense_anchor_publication_manifest()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to read staged dense anchor publication identity: {error}"
-            ))
-        })?
-        .is_none();
-    let publication = next_index_publication(
-        previous_publication.as_ref(),
-        IndexPublicationMode::Incremental,
-        &Uuid::new_v4().to_string(),
-    )?;
-    let source_identity = format!("core:{}:{}", publication.generation_id, publication.run_id);
-    preparation
-        .staged_mut()
-        .store_mut()
-        .begin_incremental_run()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to persist staged incomplete index marker: {error}"
-            ))
-        })?;
-    preparation
-        .staged_mut()
-        .store_mut()
-        .invalidate_grounding_snapshots()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to invalidate staged derived index snapshots: {error}"
-            ))
-        })?;
-    let (execution_plan, mut policy_exclusions) = incremental_execution_plan(
-        preparation.staged_mut(),
-        root,
-        storage_path,
-        source_index_policy,
-    )?;
-    let mut semantic_plan =
-        plan_incremental_semantics(preparation.staged_mut(), root, &execution_plan)?;
-    let stats = run_incremental_indexer(
-        preparation.staged_mut(),
-        root,
-        events_tx,
-        cancel_token,
-        source_index_policy,
-        &execution_plan,
-        &mut semantic_plan,
-        &mut policy_exclusions,
-    )?;
-    validate_incremental_refresh_coverage(preparation.staged_mut(), root)?;
-    let semantic_refresh_scope = incremental_semantic_refresh_scope(
-        preparation.staged_mut(),
-        root,
-        &execution_plan,
-        &semantic_plan,
-    )?;
-    let semantic_stats = finalize_staged_semantic_docs_for_runtime(
-        preparation.staged_mut().store_mut(),
-        (!rebuild_complete_dense_anchor_set).then_some(&semantic_refresh_scope),
-        (!rebuild_complete_dense_anchor_set).then_some(&semantic_plan.component_reports),
-        &source_identity,
-        cancel_token,
-        runtime,
-        SemanticProjectionDocumentSource::SourceFiles,
-    )?;
-    ensure_indexing_active(cancel_token)?;
-    let finalize_stats = preparation
-        .staged_mut()
-        .snapshots()
-        .finalize_staged()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to finalize staged incremental storage: {error}"
-            ))
-        })?;
-    let detail_started = Instant::now();
-    preparation
-        .staged_mut()
-        .snapshots()
-        .refresh_detail()
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Failed to refresh staged grounding detail snapshot: {error}"
-            ))
-        })?;
-    ensure_indexing_active(cancel_token)?;
-    Ok(PreparedIncrementalRefresh {
-        staged: preparation.release(),
-        publication,
-        stats,
-        finalize_stats,
-        detail_snapshot_ms: clamp_u128_to_u32(detail_started.elapsed().as_millis()),
-        semantic_stats,
-        semantic_refresh_scope,
-        policy_exclusions,
-    })
-}
-
-fn run_incremental_indexing_common(
-    root: &Path,
-    storage_path: &Path,
-    events_tx: &Sender<AppEventPayload>,
-    cancel_token: Option<&CancellationToken>,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-    source_index_policy: &SourceIndexPolicy,
-) -> Result<IndexingRunSummary, ApiError> {
-    let PreparedIncrementalRefresh {
-        mut staged,
-        publication,
-        stats: index_stats,
-        finalize_stats: staged_finalize_stats,
-        detail_snapshot_ms,
-        semantic_stats: staged_semantic_stats,
-        semantic_refresh_scope: llm_refresh_scope,
-        policy_exclusions,
-    } = prepare_incremental_refresh(
-        root,
-        storage_path,
-        events_tx,
-        cancel_token,
-        runtime,
-        source_index_policy,
-    )?;
-    let workspace = match runtime_workspace_manifest(root, storage_path) {
-        Ok(workspace) => workspace,
-        Err(error) => {
-            let _ = staged.discard();
-            return Err(ApiError::internal(format!(
-                "Failed to reopen project: {error}"
-            )));
-        }
-    };
-    if let Err(error) = stage_core_publication_identity(
-        &mut staged,
-        root,
-        &workspace,
-        &publication,
-        &policy_exclusions,
-        source_index_policy,
-        cancel_token,
-    ) {
-        let _ = staged.discard();
-        return Err(error);
-    }
-    let prepared_search_state = match rebuild_search_state_from_storage_for_runtime(
-        staged.store_mut(),
-        storage_path,
-        Some(&llm_refresh_scope),
-        false,
-        runtime,
-        cancel_token,
-        None,
-    ) {
-        Ok(state) => state,
-        Err(error) => {
-            let _ = staged.discard();
-            discard_unpublished_search_generation(storage_path, &publication);
-            return Err(error);
-        }
-    };
-    if is_indexing_cancelled(cancel_token) {
-        drop(prepared_search_state);
-        let _ = staged.discard();
-        discard_unpublished_search_generation(storage_path, &publication);
-        return Err(indexing_cancelled_error());
-    }
-    let prepared_commit =
-        PreparedCoreCommit::new(staged, prepared_search_state, storage_path, &publication);
-    let (prepared_search_state, staged_publish_stats, publish_duration) =
-        prepared_commit.commit(CoreCommitMode::Incremental, cancel_token)?;
-    let phase_timings = core_indexing_phase_timings(
-        &index_stats,
-        staged_finalize_stats,
-        detail_snapshot_ms,
-        staged_publish_stats,
-        publish_duration,
-        staged_semantic_stats.semantic_context_index_ms,
-    );
-    Ok(IndexingRunSummary {
-        phase_timings,
-        staged_semantic_stats,
-        llm_refresh_scope: Some(llm_refresh_scope),
-        #[cfg(test)]
-        publication,
-        prepared_search_state: Some(prepared_search_state),
-    })
-}
-
-fn is_indexing_cancelled(cancel_token: Option<&CancellationToken>) -> bool {
-    cancel_token
-        .map(CancellationToken::is_cancelled)
-        .unwrap_or(false)
-}
-
-fn indexing_cancelled_error() -> ApiError {
-    ApiError::new("cancelled", "Indexing cancelled.")
-}
-
-fn ensure_indexing_active(cancel_token: Option<&CancellationToken>) -> Result<(), ApiError> {
-    if is_indexing_cancelled(cancel_token) {
-        Err(indexing_cancelled_error())
-    } else {
-        Ok(())
-    }
-}
-
-fn workspace_refresh_inputs(store: &Store) -> Result<RefreshInputs, ApiError> {
-    Ok(RefreshInputs {
-        stored_files: store
-            .files()
-            .inventory()
-            .map_err(|e| ApiError::internal(format!("Failed to read workspace inventory: {e}")))?,
-        policy_exclusions: store
-            .get_source_policy_exclusions()
-            .map_err(|e| {
-                ApiError::internal(format!("Failed to read source policy exclusions: {e}"))
-            })?
-            .iter()
-            .map(source_policy_exclusion_candidate)
-            .collect(),
-        inventory: Default::default(),
-    })
-}
-
-fn reuse_completed_search_state(
-    storage: &mut Storage,
-    search_storage_path: &Path,
-    publication: &IndexPublicationRecord,
-    hydrate_semantic_docs: bool,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-    cancel_token: Option<&CancellationToken>,
-) -> Result<Option<SearchStateBuildResult>, ApiError> {
-    let generation_id = Uuid::parse_str(&publication.generation_id)
-        .map_err(|error| {
-            ApiError::internal(format!(
-                "Invalid index publication generation id {}: {error}",
-                publication.generation_id
-            ))
-        })?
-        .to_string();
-    let Some(marker) = read_search_generation_completion(search_storage_path, &generation_id)
-    else {
-        return Ok(None);
-    };
-
-    let search_index_started = Instant::now();
-    let mut engine = match SearchEngine::open_existing(search_storage_path) {
-        Ok(engine) => engine,
-        Err(error) => {
-            tracing::warn!(
-                path = %search_storage_path.display(),
-                "Completed persisted search generation could not be reopened and will be rebuilt: {error}"
-            );
-            return Ok(None);
-        }
-    };
-    engine.load_symbol_projection(std::iter::empty());
-    let (node_names, mut search_stats) = load_canonical_search_symbols(
-        storage,
-        SEARCH_SYMBOL_STREAM_BATCH_SIZE,
-        cancel_token,
-        |batch| {
-            engine.extend_symbol_projection(
-                batch
-                    .into_iter()
-                    .map(|entry| (entry.node_id, entry.display_name)),
-            );
-            Ok(())
-        },
-    )?;
-    if marker.symbol_count != search_stats.search_symbol_stream_rows as u64
-        || engine.full_text_doc_count() != search_stats.search_symbol_stream_rows as usize
-        || engine.tantivy_doc_count() as u64 != marker.tantivy_doc_count
-    {
-        tracing::warn!(
-            path = %search_storage_path.display(),
-            searchable_docs = engine.full_text_doc_count(),
-            stored_docs = engine.tantivy_doc_count(),
-            expected_symbols = search_stats.search_symbol_stream_rows,
-            expected_stored_docs = marker.tantivy_doc_count,
-            "Completed persisted search generation count validation failed and will be rebuilt"
-        );
-        return Ok(None);
-    }
-    search_stats.search_symbol_index_ms =
-        clamp_u128_to_u32(search_index_started.elapsed().as_millis());
-    let semantic_stats = load_persisted_semantic_docs_for_runtime(
-        storage,
-        &mut engine,
-        hydrate_semantic_docs,
-        runtime,
-    )?;
-    Ok(Some(SearchStateBuildResult {
-        publication: Some(publication.clone()),
-        node_names,
-        engine,
-        search_stats,
-        semantic_stats,
-    }))
-}
-
-fn build_persisted_search_state_from_canonical_symbols(
-    storage: &mut Storage,
-    search_storage_path: &Path,
-    hydrate_semantic_docs: bool,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-    cancel_token: Option<&CancellationToken>,
-) -> Result<SearchStateBuildResult, ApiError> {
-    let search_index_started = Instant::now();
-    let count_started = Instant::now();
-    let expected_rows = storage
-        .get_canonical_search_symbol_count()
-        .map_err(|error| {
-            ApiError::internal(format!("Failed to count canonical search symbols: {error}"))
-        })?;
-    let mut stream_duration = count_started.elapsed();
-    let mut engine = SearchEngine::new(Some(search_storage_path)).map_err(|error| {
-        if search::engine::is_persisted_search_index_busy(&error) {
-            ApiError::new(
-                "cache_busy",
-                format!("Failed to init search engine: {error}"),
-            )
-        } else {
-            ApiError::internal(format!("Failed to init search engine: {error}"))
-        }
-    })?;
-    let mut node_names = HashMap::with_capacity(expected_rows as usize);
-    let mut symbol_session = engine.begin_symbol_index().map_err(|error| {
-        ApiError::internal(format!("Failed to start symbol index writer: {error}"))
-    })?;
-    let mut after_node_id = None;
-    let mut stream_rows = 0_usize;
-    let mut stream_batches = 0_usize;
-    loop {
-        let batch_started = Instant::now();
-        let batch = storage
-            .get_canonical_search_symbol_batch_after(after_node_id, SEARCH_SYMBOL_STREAM_BATCH_SIZE)
-            .map_err(|error| {
-                ApiError::internal(format!(
-                    "Failed to stream canonical search symbols: {error}"
-                ))
-            })?;
-        stream_duration = stream_duration.saturating_add(batch_started.elapsed());
-        if batch.is_empty() {
-            break;
-        }
-        after_node_id = batch.last().map(|entry| entry.node_id);
-        stream_rows = stream_rows.saturating_add(batch.len());
-        stream_batches = stream_batches.saturating_add(1);
-        let symbols = batch
-            .into_iter()
-            .map(|entry| {
-                node_names.insert(entry.node_id, entry.display_name.clone());
-                (entry.node_id, entry.display_name)
-            })
-            .collect::<Vec<_>>();
-        symbol_session.add_nodes(symbols).map_err(|error| {
-            ApiError::internal(format!("Failed to index search nodes: {error}"))
-        })?;
-        #[cfg(test)]
-        publication_test_checkpoint(PublicationTestBoundary::SearchSymbolPage, cancel_token)?;
-        if is_indexing_cancelled(cancel_token) {
-            return Err(indexing_cancelled_error());
-        }
-    }
-    if stream_rows != expected_rows as usize {
-        return Err(ApiError::internal(format!(
-            "Canonical search symbol stream count changed: expected {expected_rows}, loaded {stream_rows}"
-        )));
-    }
-    #[cfg(test)]
-    publication_test_checkpoint(PublicationTestBoundary::SearchIndexWrite, cancel_token)?;
-    if is_indexing_cancelled(cancel_token) {
-        return Err(indexing_cancelled_error());
-    }
-    let symbol_write_stats = symbol_session
-        .finish()
-        .map_err(|error| ApiError::internal(format!("Failed to commit symbol index: {error}")))?;
-    if engine.full_text_doc_count() != stream_rows {
-        return Err(ApiError::internal(format!(
-            "Persisted search generation validation failed: indexed {} docs for {stream_rows} canonical symbols",
-            engine.full_text_doc_count()
-        )));
-    }
-    let search_symbol_index_ms = clamp_u128_to_u32(search_index_started.elapsed().as_millis());
-    let semantic_stats = load_persisted_semantic_docs_for_runtime(
-        storage,
-        &mut engine,
-        hydrate_semantic_docs,
-        runtime,
-    )?;
-    Ok(SearchStateBuildResult {
-        publication: None,
-        node_names,
-        engine,
-        search_stats: SearchStateBuildStats {
-            search_projection_rebuild_ms: 0,
-            search_symbol_stream_ms: clamp_u128_to_u32(stream_duration.as_millis()),
-            search_symbol_stream_rows: clamp_usize_to_u32(stream_rows),
-            search_symbol_stream_batches: clamp_usize_to_u32(stream_batches),
-            search_symbol_index_ms,
-            search_symbol_index_docs_written: clamp_usize_to_u32(symbol_write_stats.docs_written),
-            search_symbol_index_writer_count: clamp_usize_to_u32(symbol_write_stats.writer_count),
-            search_symbol_index_commit_count: clamp_usize_to_u32(symbol_write_stats.commit_count),
-            search_symbol_index_reload_count: clamp_usize_to_u32(symbol_write_stats.reload_count),
-            search_symbol_index_commit_ms: clamp_u128_to_u32(
-                symbol_write_stats.commit_duration.as_millis(),
-            ),
-            search_symbol_index_reload_ms: clamp_u128_to_u32(
-                symbol_write_stats.reload_duration.as_millis(),
-            ),
-        },
-        semantic_stats,
-    })
-}
-
-#[cfg(test)]
-fn rebuild_search_state_from_storage(
-    storage: &mut Storage,
-    storage_path: &Path,
-    llm_refresh_scope: Option<&HashSet<codestory_contracts::graph::NodeId>>,
-    hydrate_semantic_docs: bool,
-) -> Result<SearchStateBuildResult, ApiError> {
-    rebuild_search_state_from_storage_for_runtime(
-        storage,
-        storage_path,
-        llm_refresh_scope,
-        hydrate_semantic_docs,
-        &test_sidecar_runtime_from_env(),
-        None,
-        None,
-    )
-}
-
-type SearchCompletionValidator<'a> =
-    &'a mut dyn FnMut(&IndexPublicationRecord) -> Result<(), ApiError>;
-
-fn rebuild_search_state_from_storage_for_runtime(
-    storage: &mut Storage,
-    storage_path: &Path,
-    _llm_refresh_scope: Option<&HashSet<codestory_contracts::graph::NodeId>>,
-    hydrate_semantic_docs: bool,
-    runtime: &codestory_retrieval::SidecarRuntimeConfig,
-    cancel_token: Option<&CancellationToken>,
-    mut validate_before_completion: Option<SearchCompletionValidator<'_>>,
-) -> Result<SearchStateBuildResult, ApiError> {
-    let publication = storage.get_index_publication().map_err(|error| {
-        ApiError::internal(format!(
-            "Failed to read search publication identity: {error}"
-        ))
-    })?;
-    let _catalog_guard = publication
-        .as_ref()
-        .map(|_| SearchGenerationCatalogGuard::acquire(storage_path))
-        .transpose()?;
-    let search_storage_path =
-        search_index_path_for_publication(storage_path, publication.as_ref())?;
-    let reused = match publication.as_ref() {
-        Some(publication) => reuse_completed_search_state(
-            storage,
-            &search_storage_path,
-            publication,
-            hydrate_semantic_docs,
-            runtime,
-            cancel_token,
-        )?,
-        None => None,
-    };
-    if is_indexing_cancelled(cancel_token) {
-        return Err(indexing_cancelled_error());
-    }
-    #[cfg(test)]
-    publication_test_checkpoint(PublicationTestBoundary::SearchBuild, cancel_token)?;
-    let built_new = reused.is_none();
-    let mut result = match reused {
-        Some(result) => result,
-        None => build_persisted_search_state_from_canonical_symbols(
-            storage,
-            search_storage_path.as_path(),
-            hydrate_semantic_docs,
-            runtime,
-            cancel_token,
-        )
-        .map_err(|mut error| {
-            error.message = format!("Failed to rebuild search state: {}", error.message);
-            error
-        })?,
-    };
-    if is_indexing_cancelled(cancel_token) {
-        return Err(indexing_cancelled_error());
-    }
-    #[cfg(test)]
-    publication_test_checkpoint(PublicationTestBoundary::SearchValidation, cancel_token)?;
-    if result.engine.full_text_doc_count() != result.node_names.len() {
-        return Err(ApiError::internal(format!(
-            "Prepared search generation contains {} searchable symbols for {} core symbols",
-            result.engine.full_text_doc_count(),
-            result.node_names.len()
-        )));
-    }
-    if built_new && let Some(publication) = publication.as_ref() {
-        if is_indexing_cancelled(cancel_token) {
-            return Err(indexing_cancelled_error());
-        }
-        if let Some(validate) = validate_before_completion.as_mut()
-            && let Err(error) = validate(publication)
-        {
-            drop(result);
-            discard_unpublished_search_generation(storage_path, publication);
-            return Err(error);
-        }
-        #[cfg(test)]
-        publication_test_checkpoint(PublicationTestBoundary::SearchCompletion, cancel_token)?;
-        write_search_generation_completion(
-            &search_storage_path,
-            publication,
-            result.node_names.len(),
-            result.engine.tantivy_doc_count(),
-        )?;
-    }
-    if publication.is_some() {
-        result
-            .engine
-            .downgrade_persisted_lock_to_shared()
-            .map_err(|error| {
-                ApiError::internal(format!(
-                    "Failed to share completed search generation {}: {error}",
-                    search_storage_path.display()
-                ))
-            })?;
-    }
-    if publication.is_some()
-        && let Some(active_generation_id) =
-            search_storage_path.file_name().and_then(|id| id.to_str())
-        && let Err(error) = prune_search_generations(storage_path, active_generation_id)
-    {
-        tracing::warn!(
-            generation_id = %active_generation_id,
-            "Failed to prune persisted search generations after publication: {}",
-            error.message
-        );
-    }
-    result.publication = publication;
-    Ok(result)
-}
-
-fn refresh_caches(
-    controller: &AppController,
-    storage: &mut Storage,
-    storage_path: &Path,
-    llm_refresh_scope: Option<&HashSet<codestory_contracts::graph::NodeId>>,
-) -> Result<CacheRefreshStats, ApiError> {
-    let refreshed = rebuild_search_state_from_storage_for_runtime(
-        storage,
-        storage_path,
-        llm_refresh_scope,
-        true,
-        &controller.runtime_config,
-        None,
-        None,
-    );
-
-    match refreshed {
-        Ok(result) => Ok(publish_prepared_search_state(controller, result)),
-        Err(error) => {
-            tracing::warn!(
-                "Failed to rebuild search caches from storage: {}",
-                error.message
-            );
-            let mut state = controller.state.lock();
-            state.node_names.clear();
-            clear_search_engine(&mut state);
-            controller.sidecar_query_cache.lock().clear();
-            state.is_indexing = false;
-            Err(error)
-        }
-    }
-}
-
-fn publish_prepared_search_state(
-    controller: &AppController,
-    result: SearchStateBuildResult,
-) -> CacheRefreshStats {
-    let publish_started = Instant::now();
-    let mut state = controller.state.lock();
-    state.node_names = result.node_names;
-    publish_search_engine(&mut state, result.engine, result.publication);
-    controller.sidecar_query_cache.lock().clear();
-    state.is_indexing = false;
-    CacheRefreshStats {
-        search_stats: result.search_stats,
-        semantic_stats: result.semantic_stats,
-        runtime_cache_publish_ms: Some(clamp_u128_to_u32(publish_started.elapsed().as_millis())),
-    }
-}
-
-#[cfg(test)]
-mod tests;
